% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BotanicalCorrection.R
\name{BotanicalCorrection}
\alias{BotanicalCorrection}
\title{Botanical Correction}
\usage{
BotanicalCorrection(Data, Source = c("TPL", "WFO"), WFOData = NULL)
}
\arguments{
\item{Data}{Dataset (data.frame or data.table)
The dataset must contain the columns:
\itemize{
\item \code{IdTree} (character)
\item \code{Family} (character)
\item \code{Genus} (character)
\item \code{Species} (character)
\item \code{VernName} (character)
\item \code{ScientificName} (character)
}}

\item{Source}{(character, 1 value) To correct and standardise, you can choose between:
\itemize{
\item "TPL": \emph{The Plant List} (http://www.theplantlist.org/) (faster but based
on the 2013 taxonomy)
\item "WFO": \emph{World Flora Online} (http://www.worldfloraonline.org/) (long time
but based on the 2022 taxonomy)
}}

\item{WFOData}{If \code{Source} = "WFO".
Data setwith the static copy of the \emph{World Flora Online} (WFO) Taxonomic Backbone
data (from http://www.worldfloraonline.org/downloadData).
data.frame or data.table in R, .rds file if you are uploading from the Shiny App.
Note that the classification.csv file downloaded from the website it too large to be uploaded
in the App, so it needs to be opened in R first, trimmed to the set of species
relevant to you, and saved to an .rds file using function saveRDS().}
}
\value{
Fill the \emph{Comment_TreeData} column with error type informations and add columns:
\itemize{
\item \code{Family_TreeDataCor} (character): corrected Family name
\item \code{FamilyCorSource} (character): source of the Family correction
\item \code{Genus_TreeDataCor} (character): corrected Genus name
\item \code{Species_TreeDataCor} (character): corrected Species name
\item \code{BotanicalCorrectionSource} (character): source of the Genus and Species
correction
\item \code{ScientificName_TreeDataCor} (character): corrected Scientific name
\item \code{VernName_TreeDataCor} (character): completed if information available at
\code{IdTree} level.
}
}
\description{
Botanical Correction
}
\details{
\itemize{
\item No special characters (typography)
\item No family name in the Genus and Species columns (the suffix "aceae" is
specific to the family name.
\item Correct spelling of botanical names (\emph{Taxonstand or WorldFlora})
\item Family & Scientific names match (\emph{BIOMASS::getTaxonomy or WorldFlora})
\item Update the scientific botanical names with the current phylogenetic
classification
\item Check \strong{invariant botanical informations per IdTree} (1 IdTree = 1 family,
1 scientific and 1 vernacular name)
}
}
\examples{
\dontrun{
library(data.table)
data(TestData)

# With The Plant List:
Rslt <- BotanicalCorrection(TestData, Source = "TPL")

ScfcCor <- unique(Rslt[ScientificNameCor != ScientificName,
                list(ScientificName, ScientificNameCor,
                Family, FamilyCor, FamilyCorSource,
                Genus, GenusCor,
                Species, SpeciesCor, Subspecies,
                BotanicalCorrectionSource, Comment)
                ])

FamCor <- unique(Rslt[FamilyCor != Family,
                list(ScientificName, ScientificNameCor,
                Family, FamilyCor, FamilyCorSource,
                Genus, GenusCor,
                Species, SpeciesCor, Subspecies,
                BotanicalCorrectionSource, Comment)
                ])

# With World Flora Online:
WFO_Backbone <- file.choose()
load(WFO_Backbone)

RsltWFO <- BotanicalCorrection(TestData, Source = "WFO", WFOData = WFO_Backbone)

ScfcCor <- unique(RsltWFO[ScientificNameCor != ScientificName,
                list(ScientificName, ScientificNameCor,
                Family, FamilyCor,
                Genus, GenusCor,
                Species, SpeciesCor, Subspecies,
                BotanicalCorrectionSource, Comment)
                ])

FamCor <- unique(RsltWFO[FamilyCor != Family,
                list(ScientificName, ScientificNameCor,
                Family, FamilyCor,
                Genus, GenusCor,
                Species, SpeciesCor, Subspecies,
                BotanicalCorrectionSource, Comment)
                ])
}

}
