% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeMeasuredValues.R
\name{ComputeMeasuredValues}
\alias{ComputeMeasuredValues}
\title{Compute Measured Values}
\usage{
ComputeMeasuredValues(initial_values, cresc_abs)
}
\arguments{
\item{initial_values}{Vector of the initial values of the variable (numeric)}

\item{cresc_abs}{Vector of the absolute incrementations of the variable (numeric)}
}
\value{
A numeric vector of the same size as \emph{initial_values} with, the first
initial value incremented by the first value of cresc_abs, then this value
incremented by the 2nd value of cresc_abs, then this value incremented by
the 3rd value of cresc_abs, etc.
}
\description{
Compute Measured Values
}
\examples{
initial_values <- c(3.7, 2.8, 2.8, 2.8, NA)
cresc_abs <- c(-0.9, 0.0,  0.0)
ComputeMeasuredValues(initial_values = initial_values,
                      cresc_abs = cresc_abs)

}
