% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiameterCorrection.R
\name{DiameterCorrectionByTree}
\alias{DiameterCorrectionByTree}
\title{DiameterCorrectionByTree}
\usage{
DiameterCorrectionByTree(
  DataTree,
  Data,
  DefaultHOM = 1.3,
  MaxDBH = 500,
  PositiveGrowthThreshold = 5,
  NegativeGrowthThreshold = -2,
  Pioneers = NULL,
  PioneersGrowthThreshold = 7.5,
  TrustMeasSet = "first",
  WhatToCorrect = c("POM change", "punctual", "shift"),
  CorrectionType = c("quadratic", "linear", "individual", "phylogenetic hierarchical"),
  DBHRange = 10,
  MinIndividualNbr = 5,
  Digits = 1L,
  DetectOnly = FALSE
)
}
\arguments{
\item{DataTree}{A dataset corresponding to a single tree/stem's (1 IdTree/IdStem)
measurements (data.frame or data.table).}

\item{Data}{Complete dataset (data.table) used if the "phylogenetic
hierarchical" correction (\emph{CorrectionType} argument) is chosen.
The dataset must contain the columns:
\itemize{
\item \code{IdTree} (character)
\item \code{IdStem} (character) if \emph{ByStem} argument = TRUE
\item \code{ScientificName} (character)
\item \code{Genus} (character)
\item \code{Family} (character)
\item \code{Diameter} (numeric)
\item \code{Year} (numeric)
}}

\item{DefaultHOM}{Default Height Of Measurement in meter (Default: 1.3 m)
(numeric, 1 value)}

\item{MaxDBH}{Maximum possible DBH of your stand in cm (numeric, 1 value)}

\item{PositiveGrowthThreshold}{in cm/year: a tree
widening by more than x cm/year is considered abnormal (numeric, 1 value)}

\item{NegativeGrowthThreshold}{in cm/census: the possible
positive measurement error (+n) cannot be corrected until the growth
appears abnormal, but a negative measurement error can be allowed until -n
(a tree does not decrease). Thus the positive measurement error (+n) is
"compensated". (numeric, 1 value)}

\item{Pioneers}{Scientific names of the pioneer species of the site, as in
the 'ScientificName' column (characters vector)}

\item{PioneersGrowthThreshold}{in cm/year: a tree of a pioneer species that
widens by more than x cm/year is considered abnormal (numeric, 1 value)}

\item{TrustMeasSet}{Trust measurements set: the "first" or the "last" set
(character, 1 value)}

\item{WhatToCorrect}{c("POM change", "punctual", "shift") (character)
\itemize{
\item "POM change": detect POM change in the column 'POM' and correct
the Diameter values from it.
\item "punctual": detect if the error is punctual and correct it by
interpolation.
\item "shift": detect if there is a shift of several 'Diameter' values and
links them to the trust measurements set
(\emph{TrustMeasSet} argument).
}}

\item{CorrectionType}{c("linear", "quadratic", "individual", "phylogenetic
hierarchical") (character).
\itemize{
\item "linear": interpolation by linear regression of the individual annual
growth over time.
\item "quadratic": interpolation by quadratic regression  of the individual
annual growth over time.
\item "individual": replace abnormal growth by interpolation from the
individual values.
\item "phylogenetic hierarchical": replace abnormal growth with the average
growth of other trees in the dataset, at the specific, genus, family
or stand level, within a DBH range of x cm (\emph{DBHRange} argument).
If the number of these trees < n (\emph{MinIndividualNbr} argument)
at the specific level, we switch to the genus level etc.
}}

\item{DBHRange}{DBH range in cm to take into account to select other trees in
the dataset to apply "phylogenetic hierarchical" correction (Default: 10
cm) (numeric, 1 value)}

\item{MinIndividualNbr}{Minimum number of individuals to take into account in
"phylogenetic hierarchical" correction (Default: 5) (numeric, 1 value)}

\item{Digits}{Number of decimal places to be used in the 'DBHCor' column
(Default: 1L) (integer)}

\item{DetectOnly}{TRUE: Only detect errors, FALSE: detect and correct errors
(Default: FALSE) (logical)}
}
\value{
Fill the \emph{Comment} column with error type informations. If
\emph{DetectOnly} = FALSE, add columns:
\itemize{
\item \emph{DBHCor}: corrected trees diameter at default HOM
\item \emph{DiameterCorrectionMeth} = "linear"/"quadratic"/"individual"
/phylogenetic hierarchical("species"/"genus"/"family"/"stand"/"shift
realignment")
}
}
\description{
DiameterCorrectionByTree
}
\details{
When there is only 1 \code{Diameter} value for a tree/stem, \code{DBHCor}
takes the original \code{Diameter} value. If this value is 0 or > MaxDBH,
\code{DBHCor} takes NA.
}
\examples{
library(data.table)
data(TestData)

 DataTree <- data.table(IdTree = "c",
      ScientificName = "A",
      Year = c(seq(2000,2008, by = 2), 2012, 2014,2016, 2020), # 9 Diameter values
      Diameter = c(13:16, 16-4, (16-4)+2, (16-4)+3, 15-4, (15-4)+2), # 0.5 cm/year
      POM = as.factor(c(0, 0, 0, 0, 1, 1, 1, 2, 2)))

Rslt <- DiameterCorrectionByTree(
  DataTree, TestData,
  WhatToCorrect = c("POM change", "punctual", "shift"),
  CorrectionType = c("quadratic", "linear", "individual")
  )

}
