% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiameterCorrectionPlot.R
\name{DiameterCorrectionPlot}
\alias{DiameterCorrectionPlot}
\title{Diameter Correction Plot}
\usage{
DiameterCorrectionPlot(Data)
}
\arguments{
\item{Data}{Dataset (data.frame or data.table)
The dataset must contain the columns:
\itemize{
\item \code{IdStem} (character)
\item \code{Year} (numeric)
\item \code{Diameter} (numeric)
\item \code{DBHCor} (numeric)
\code{HOM} (Height Of Measurement) (numeric)
\code{HOMCor} (Corrected Height Of Measurement) (numeric)
}}
}
\value{
Saves a pdf to the root of your R project with the plots of the
initial diameter values and proposed corrections, by IdStem.
}
\description{
Diameter Correction Plot
}
\examples{

# DiameterCorrectionPlot(Rslt)

}
