% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErrorsDetection.R
\name{ErrorsDetection}
\alias{ErrorsDetection}
\title{ErrorsDetection}
\usage{
ErrorsDetection(
  Data,
  DeathConfirmation = 2,
  UseSize = FALSE,
  MinDBH = 10,
  PositiveGrowthThreshold = 5
)
}
\arguments{
\item{Data}{(data.frame or data.table)}

\item{DeathConfirmation}{Number of times (censuses) needed for an unseen tree
to be considered dead (numeric) (Default = 2 censuses)}

\item{UseSize}{Use the size presence as a witness of the living status of the
tree (logical) (Default = FALSE)}

\item{MinDBH}{Minimum diameter of trees inventoried according to your
protocol (in cm) (numeric, 1 value) (Default = 10 cm)}

\item{PositiveGrowthThreshold}{A tree widening by more than x cm/year is
considered abnormal (numeric, 1 value) (Default = 5 cm)}
}
\value{
The input dataset (data.table) with a new \emph{Comment} column with error
type informations.
}
\description{
ErrorsDetection
}
\details{
Detect errors
\itemize{
\item Remove \emph{duplicated rows}
\item Check \emph{missing value} in
X-Yutm/PlotArea/Plot/SubPlot/Year/TreeFieldNum/
IdTree/DBH/MeasCode/Family/Genus/Species/VernName
\item Check \emph{missing value} (NA/0) in the measurement variables
\item Check \emph{duplicated TreeFieldNum} in plot-subplot association in a census
(at the site scale)
\item Check of the \emph{unique association of the idTree with plot, subplot and
TreeFieldNum} (at the site scale)
\item Check \emph{duplicated idTree} in a census (at the site scale)
\item Check for trees \emph{outside the subplot}
\item Check \emph{invariant coordinates per IdTree}
\item Check \emph{fix Plot and SubPlot number}
\item Check botanical identification (BotanicalCorrection)
\item Check the life status evolution of the trees (StatusCorrection)
\item Check diameter evolution of the trees (DiameterCorrection)
\item Check recruitment (RecruitmentCorrection)
}
}
\examples{
library(data.table)
data("TestData")

Rslt <- ErrorsDetection(TestData)

}
\seealso{
\code{\link{StatusCorrection}}
}
