% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleForestGEO.R
\docType{data}
\name{ExampleForestGEO}
\alias{ExampleForestGEO}
\title{ExampleForestGEO}
\format{
See examples.
\describe{
#' To interpret the names of the datasets, see * \url{http://ctfs.si.edu/Public/CTFSRPackage/files/help/RoutputStem.pdf} and * \url{https://raw.githubusercontent.com/forestgeo/fgeo.data/master/man/census_description.Rd}
...
}
}
\usage{
ExampleForestGEO
}
\description{
Stem tables from Luquillo produced with
\href{https://github.com/forestgeo/rtbl/releases/tag/0.0.0.9000}{\strong{rtbl} 0.0.0.9000}, with Species table merge into it and some alteration of dbh for troubleshooting purpous. NOT SUITABLE FOR DATA ANALYSIS.
}
\details{
For a general description see \link{census_description}.
}
\seealso{
\itemize{
\item \url{https://forestgeo.si.edu/sites/north-america/luquillo}.
}
}
\keyword{datasets}
