% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatusCorrection.R
\name{FillinInvariantColumns}
\alias{FillinInvariantColumns}
\title{FillinInvariantColumns}
\usage{
FillinInvariantColumns(NewRow, InvariantColumns, DataTree, IdTree)
}
\arguments{
\item{NewRow}{The dataset to fill (data.table)}

\item{InvariantColumns}{Vector with the names of the columns that are
supposed to have always the same value for each measurement of the same
tree (character)}

\item{DataTree}{A dataset corresponding to a single tree's (1 IdTree)
measurements, with the invariant columns and their value (data.table)}

\item{IdTree}{(character)}
}
\value{
Fill the invariant columns in NewRow with their (unique) value
}
\description{
Fill the invariant columns in NewRow with their (unique) value
}
