% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForestGEOProfile.R
\docType{data}
\name{ForestGEOProfile}
\alias{ForestGEOProfile}
\title{ForestGEOProfile}
\format{
A named list
\describe{
\item{TreeHeightUnitMan}{Value or column name in data set @ForestGEOProfile (none) corresponding to What is your tree height unit?}
\item{Family}{Value or column name in data set @ForestGEOProfile (Family) corresponding to Family name}
\item{TreeHeight}{Value or column name in data set @ForestGEOProfile (none) corresponding to Tree height}
\item{Date}{Value or column name in data set @ForestGEOProfile (ExactDate) corresponding to Date of measurement}
\item{POM}{Value or column name in data set @ForestGEOProfile (hom) corresponding to POM}
\item{PlotAreaUnitMan}{Value or column name in data set @ForestGEOProfile (none) corresponding to What is the unit for plot area?}
\item{Xutm}{Value or column name in data set @ForestGEOProfile (none) corresponding to Tree X UTM coordinate}
\item{IdStem}{Value or column name in data set @ForestGEOProfile (stemID) corresponding to Stem ID}
\item{PlotMan}{Value or column name in data set @ForestGEOProfile (ForestGEO) corresponding to What is your plot name ?}
\item{Species}{Value or column name in data set @ForestGEOProfile (Species) corresponding to Species name}
\item{ScientificNameSep}{Value or column name in data set @ForestGEOProfile () corresponding to Character separating Genus and Species in Scientific name}
\item{Yfield}{Value or column name in data set @ForestGEOProfile (gy) corresponding to Tree Y euclidean position in plot}
\item{CircUnitMan}{Value or column name in data set @ForestGEOProfile (none) corresponding to What is your circumference unit?}
\item{IdMeasure}{Value or column name in data set @ForestGEOProfile (MeasureID) corresponding to Measure ID}
\item{IdLevel}{Value or column name in data set @ForestGEOProfile (IDLevel) corresponding to Deepest taxonomic level for which full identification is known}
\item{TreeFieldNum}{Value or column name in data set @ForestGEOProfile (tag) corresponding to Tree unique identifiers}
\item{ScientificName}{Value or column name in data set @ForestGEOProfile (Latin) corresponding to Scientific name}
\item{DateFormat}{Value or column name in data set @ForestGEOProfile (mm/dd/yyyy) corresponding to What is your date format}
\item{IdCensus}{Value or column name in data set @ForestGEOProfile (CensusID) corresponding to Census ID}
\item{PlotAreaMan}{Value or column name in data set @ForestGEOProfile (25.6) corresponding to What is your plot area in hectar?}
\item{Year}{Value or column name in data set @ForestGEOProfile (none) corresponding to Year of measurement}
\item{LifeStatus}{Value or column name in data set @ForestGEOProfile (DFstatus) corresponding to Life status}
\item{PlotArea}{Value or column name in data set @ForestGEOProfile (none) corresponding to Plot area}
\item{CommercialSp}{Value or column name in data set @ForestGEOProfile (none) corresponding to Commercial species indicator}
\item{SubPlotAreaUnitMan}{Value or column name in data set @ForestGEOProfile (none) corresponding to What is the unit for subplot area?}
\item{YearMan}{Value or column name in data set @ForestGEOProfile (NA) corresponding to What is your year of measurement?}
\item{Yutm}{Value or column name in data set @ForestGEOProfile (none) corresponding to Tree Y UTM coordinate}
\item{Circ}{Value or column name in data set @ForestGEOProfile (none) corresponding to Circumference}
\item{SubPlotAreaMan}{Value or column name in data set @ForestGEOProfile (0.04) corresponding to What is your subplot area in hectar?}
\item{SubPlotMan}{Value or column name in data set @ForestGEOProfile () corresponding to What is your subplot name ?}
\item{Site}{Value or column name in data set @ForestGEOProfile (none) corresponding to Site name}
\item{Lon}{Value or column name in data set @ForestGEOProfile (none) corresponding to Tree longitude}
\item{Plot}{Value or column name in data set @ForestGEOProfile (none) corresponding to Plot name}
\item{IdTree}{Value or column name in data set @ForestGEOProfile (treeID) corresponding to Tree ID}
\item{LifeForm}{Value or column name in data set @ForestGEOProfile (none) corresponding to Life form}
\item{VernName}{Value or column name in data set @ForestGEOProfile (none) corresponding to Vernacular name}
\item{SubPlotArea}{Value or column name in data set @ForestGEOProfile (none) corresponding to SubPlot area}
\item{SiteMan}{Value or column name in data set @ForestGEOProfile (SCBI) corresponding to What is your site name ?}
\item{IsLive}{Value or column name in data set @ForestGEOProfile (c("alive", "broken below")) corresponding to Which of your status(es) represent a LIVE tree?}
\item{Authority}{Value or column name in data set @ForestGEOProfile (SpeciesID) corresponding to  Taxonomic authority for the ID level}
\item{Lat}{Value or column name in data set @ForestGEOProfile (none) corresponding to Tree latitude}
\item{Genus}{Value or column name in data set @ForestGEOProfile (Genus) corresponding to Genus name}
\item{SubPlot}{Value or column name in data set @ForestGEOProfile (quadrat) corresponding to SubPlot name}
\item{DBHUnitMan}{Value or column name in data set @ForestGEOProfile (mm) corresponding to What is your DBH unit ?}
\item{Xfield}{Value or column name in data set @ForestGEOProfile (gx) corresponding to Tree X euclidean position in plot}
\item{DBH}{Value or column name in data set @ForestGEOProfile (dbh) corresponding to DBH}
\item{POMUnitMan}{Value or column name in data set @ForestGEOProfile (m) corresponding to What is your POM unit?}
...
}
}
\usage{
ForestGEOProfile
}
\description{
Generated by interacting with app or function RequiredFormat_interactive
}
\keyword{datasets}
