% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateComment.R
\name{GenerateComment}
\alias{GenerateComment}
\title{GenerateComment}
\usage{
GenerateComment(Data, condition, comment)
}
\arguments{
\item{Data}{(data.table)}

\item{condition}{(logical)}

\item{comment}{The string to add in the "Comment" column (character)}
}
\value{
The input data.table with the filled "Comment" column.
}
\description{
For all rows concerned by the condition, the function add a
string in the "Comment" column of a data.table. If there is already a
value, the string is pasted after the "/" separator.
}
\examples{
\dontrun{
library(data.table)
dt <- data.table(A = c(2, 1, 3), B = c(6, 10, 6))

dt[A == 2,
   "Comment" := paste0("A = 2")] # 1st comment

GenerateComment(dt, condition = dt$B == 6, comment = "B = 6")
}
}
