% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LifeStatusCorrectionPlot.R
\name{LifeStatusCorrectionPlot}
\alias{LifeStatusCorrectionPlot}
\title{Plot life status correction result}
\usage{
LifeStatusCorrectionPlot(Data, OnlyCorrected = FALSE, SeveralWindows = TRUE)
}
\arguments{
\item{Data}{Dataset (data.frame or data.table)
The dataset must contain the columns:
\itemize{
\item \code{IdTree} (character)
\item \code{Year} (numeric)
\item \code{LifeStatus} (logical or character)
\item \code{LifeStatusCor} (logical or character)
}}

\item{OnlyCorrected}{TRUE: plot only corrected trees, FALSE: plot all trees}
}
\value{
The plots of the initial diameter values and proposed corrections, by
IdStem.
}
\description{
Plot life status correction result
}
\examples{

\dontrun{
pdf("LifeStatusCorrectionPlots.pdf", width = 25, height = 10)
LifeStatusCorrectionPlot(Rslt, SeveralWindows = FALSE)
dev.off()
}

}
