% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParacouProfile.R
\docType{data}
\name{ParacouProfile}
\alias{ParacouProfile}
\title{ParacouProfile}
\format{
A named list
\describe{
\item{Authority}{Value or column name in data set @ParacouSubset (none) corresponding to  Taxonomic authority for the ID level}
\item{BCirc}{Value or column name in data set @ParacouSubset (none) corresponding to Basal Circumference}
\item{BCircUnitMan}{Value or column name in data set @ParacouSubset (none) corresponding to What is your basal circumference unit?}
\item{BD}{Value or column name in data set @ParacouSubset (none) corresponding to BD}
\item{BDUnitMan}{Value or column name in data set @ParacouSubset (none) corresponding to What is your BD unit ?}
\item{Circ}{Value or column name in data set @ParacouSubset (Circ) corresponding to Circumference}
\item{CircUnitMan}{Value or column name in data set @ParacouSubset (cm) corresponding to What is your circumference unit?}
\item{ClearValueName}{Some value entered via interaction with the Shiny app}
\item{CommercialSp}{Value or column name in data set @ParacouSubset (CommercialSp) corresponding to Commercial species indicator}
\item{Date}{Value or column name in data set @ParacouSubset (CensusDate) corresponding to Date of measurement}
\item{Month}{Value or column name in data set @ParacouSubset (none) corresponding to Month of measurement}
\item{Day}{Value or column name in data set @ParacouSubset (none) corresponding to Day of measurement}
\item{DateFormatMan}{Value or column name in data set @ParacouSubset (yyyy-mm-dd) corresponding to What is your date format?}
\item{Diameter}{Value or column name in data set @ParacouSubset (none) corresponding to DBH}
\item{DiameterUnitMan}{Value or column name in data set @ParacouSubset (none) corresponding to What is your DBH unit ?}
\item{Family}{Value or column name in data set @ParacouSubset (Family) corresponding to Family name}
\item{Genus}{Value or column name in data set @ParacouSubset (Genus) corresponding to Genus name}
\item{HOM}{Value or column name in data set @ParacouSubset (none) corresponding to HOM}
\item{HOMUnitMan}{Value or column name in data set @ParacouSubset (none) corresponding to What is your HOM unit?}
\item{IdCensus}{Value or column name in data set @ParacouSubset (none) corresponding to Census ID}
\item{IdLevel}{Value or column name in data set @ParacouSubset (none) corresponding to Deepest taxonomic level for which full identification is known}
\item{IdMeasure}{Value or column name in data set @ParacouSubset (none) corresponding to Measure ID}
\item{IdStem}{Value or column name in data set @ParacouSubset (none) corresponding to Stem ID}
\item{IdTree}{Value or column name in data set @ParacouSubset (idTree) corresponding to Tree ID}
\item{IsCommercialMan}{Value or column name in data set @ParacouSubset (TRUE) corresponding to Which of the following(s) indicate a commercial species?}
\item{IsLive}{Some value entered via interaction with the Shiny app}
\item{TreeLat}{Value or column name in data set @ParacouSubset (Lat) corresponding to Tree latitude}
\item{LifeForm}{Value or column name in data set @ParacouSubset (none) corresponding to Life form}
\item{LifeStatus}{Value or column name in data set @ParacouSubset (CodeAlive) corresponding to Life status}
\item{TreeLon}{Value or column name in data set @ParacouSubset (Lon) corresponding to Tree longitude}
\item{MinDBH}{Value or column name in data set @ParacouSubset (none) corresponding to Minimum DBH}
\item{MinDBHMan}{Value or column name in data set @ParacouSubset (-999) corresponding to What is the minimum diameter of trees included in your inventory?}
\item{MinDBHUnitMan}{Value or column name in data set @ParacouSubset (none) corresponding to What is your unit for the column indicating the minimum DBH? If not provided, it will be calculated looking at your Diameter column}
\item{Plot}{Value or column name in data set @ParacouSubset (Plot) corresponding to Plot name}
\item{PlotArea}{Value or column name in data set @ParacouSubset (PlotArea) corresponding to Plot area}
\item{PlotAreaMan}{Value or column name in data set @ParacouSubset (-999) corresponding to What is your plot area in hectare?}
\item{PlotAreaUnitMan}{Value or column name in data set @ParacouSubset (ha) corresponding to What is the unit for plot area?}
\item{PlotMan}{Value or column name in data set @ParacouSubset () corresponding to What is your plot name ?}
\item{TreePlotUnitMan}{Value or column name in data set @ParacouSubset (m) corresponding to Tree XY plot coordinates units}
\item{POM}{Value or column name in data set @ParacouSubset (MeasCode) corresponding to POM}
\item{ScientificName}{Value or column name in data set @ParacouSubset (none) corresponding to Scientific name}
\item{ScientificNameSepMan}{Value or column name in data set @ParacouSubset () corresponding to Character separating Genus and Species in Scientific name}
\item{Site}{Value or column name in data set @ParacouSubset (Forest) corresponding to Site name}
\item{SiteMan}{Value or column name in data set @ParacouSubset () corresponding to What is your site name ?}
\item{Species}{Value or column name in data set @ParacouSubset (Species) corresponding to Species name}
\item{SpeciesDensity}{Value or column name in data set @ParacouSubset (none) corresponding to Species density}
\item{SpeciesDensityunitMan}{Some value entered via interaction with the Shiny app}
\item{SpeciesBasalAreaDensity}{Value or column name in data set @ParacouSubset (none) corresponding to Species basal area density}
\item{SpeciesBasalAreaDensityUnitMan}{Value or column name in data set @ParacouSubset (none) corresponding to What is your species basal area unit?}
\item{SpeciesTotalBasalArea}{Value or column name in data set @ParacouSubset (none) corresponding to Species total basal area}
\item{SpeciesTotalBasalAreaUnitMan}{Value or column name in data set @ParacouSubset (none) corresponding to What is yourindividual basal area unit?}
\item{SpeciesVolume}{Value or column name in data set @ParacouSubset (none) corresponding to Species volume}
\item{SpeciesVolumeUnitMan}{Value or column name in data set @ParacouSubset (none) corresponding to What is your species volume unit?}
\item{Subplot}{Value or column name in data set @ParacouSubset (SubPlot) corresponding to Subplot name}
\item{SubplotArea}{Value or column name in data set @ParacouSubset (none) corresponding to Subplot area}
\item{SubplotAreaMan}{Value or column name in data set @ParacouSubset (1) corresponding to What is your Subplot area in hectare?}
\item{SubplotAreaUnitMan}{Value or column name in data set @ParacouSubset (none) corresponding to What is the unit for Subplot area?}
\item{SubplotMan}{Value or column name in data set @ParacouSubset () corresponding to What is your Subplot name ?}
\item{TreeSubplotUnitMan}{Value or column name in data set @ParacouSubset (none) corresponding to Tree XY subplot coordinates units}
\item{TickedMelt1}{Some value entered via interaction with the Shiny app}
\item{TickedMelt2}{Some value entered via interaction with the Shiny app}
\item{TickedMelt3}{Some value entered via interaction with the Shiny app}
\item{Tidy}{Some value entered via interaction with the Shiny app}
\item{TreeCodes}{Value or column name in data set @ParacouSubset (none) corresponding to Select all the columns that give information about your tree, and/or your methods of measurements}
\item{TreeFieldNum}{Value or column name in data set @ParacouSubset (TreeFieldNum) corresponding to Tree tag}
\item{TreeHeight}{Value or column name in data set @ParacouSubset (none) corresponding to Tree height}
\item{TreeHeightUnitMan}{Value or column name in data set @ParacouSubset (none) corresponding to What is your tree height unit?}
\item{TreeUTMUnitMan}{Value or column name in data set @ParacouSubset (m) corresponding to Tree UTM units}
\item{ValueName1}{Some value entered via interaction with the Shiny app}
\item{ValueName2}{Some value entered via interaction with the Shiny app}
\item{ValueName3}{Some value entered via interaction with the Shiny app}
\item{Variablecolumns1}{Some value entered via interaction with the Shiny app}
\item{Variablecolumns2}{Some value entered via interaction with the Shiny app}
\item{Variablecolumns3}{Some value entered via interaction with the Shiny app}
\item{VernName}{Value or column name in data set @ParacouSubset (VernName) corresponding to Vernacular name}
\item{XTreePlot}{Value or column name in data set @ParacouSubset (Xfield) corresponding to Tree X euclidean position in plot}
\item{XTreeSubplot}{Value or column name in data set @ParacouSubset (none) corresponding to Tree X euclidean position in SUBplot}
\item{XTreeUTM}{Value or column name in data set @ParacouSubset (Xutm) corresponding to Tree X UTM coordinate}
\item{Year}{Value or column name in data set @ParacouSubset (CensusYear) corresponding to Year of measurement}
\item{YearMan}{Value or column name in data set @ParacouSubset (-999) corresponding to What is your year of measurement?}
\item{YTreePlot}{Value or column name in data set @ParacouSubset (Yfield) corresponding to Tree Y euclidean position in plot}
\item{YTreeSubplot}{Value or column name in data set @ParacouSubset (none) corresponding to Tree Y euclidean position in SUBplot}
\item{YTreeUTM}{Value or column name in data set @ParacouSubset (Yutm) corresponding to Tree Y UTM coordinate}
...
}
}
\usage{
ParacouProfile
}
\description{
Generated by interacting with app or function RequiredFormat_interactive
}
\keyword{datasets}
