% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhylogeneticHierarchicalCorrection.R
\name{PhylogeneticHierarchicalCorrection}
\alias{PhylogeneticHierarchicalCorrection}
\title{PhylogeneticHierarchicalCorrection}
\usage{
PhylogeneticHierarchicalCorrection(
  DataTree,
  Data,
  cresc,
  cresc_abs,
  cresc_abn,
  DBHCor,
  Time,
  PositiveGrowthThreshold,
  NegativeGrowthThreshold,
  DBHRange = 10,
  MinIndividualNbr = 5
)
}
\arguments{
\item{DataTree}{A dataset corresponding to a single tree's (1 IdTree)
measurements (data.table)}

\item{Data}{(data.table)
The dataset must contain the columns:
\itemize{
\item 'IdTree' (character)
\item 'DBH' (numeric)
\item 'Year' (numeric)
}}

\item{cresc}{Annual diameter increment (numeric)}

\item{cresc_abs}{Absolute diameter increment (not divided by time between 2
values) (numeric)}

\item{cresc_abn}{Abnormal diameter increment positions (numeric)}

\item{DBHCor}{Diameter vector in cm (numeric)}

\item{Time}{Time variable in years (numeric)}

\item{PositiveGrowthThreshold}{in cm/year : a tree
widening by more than x cm/year is considered abnormal (numeric, 1 value)}

\item{NegativeGrowthThreshold}{in cm/census : The possible
positive measurement error (+n) cannot be corrected until the growth
appears abnormal, but a negative measurement error can be allowed until -n
(a tree does not decrease). Thus the positive measurement error (+n) is
"compensated". (numeric, 1 value)}

\item{DBHRange}{DBH range in cm to take into account to select other trees in
the dataset to apply "phylogenetic hierarchical" correction (Default: 10
cm) (numeric, 1 value)}

\item{MinIndividualNbr}{Minimum number of individuals to take into account in
"phylogenetic hierarchical" correction (Default: 5) (numeric, 1 value)}
}
\value{

}
\description{
PhylogeneticHierarchicalCorrection
}
\examples{
data(TestData)
DataTree <- TestData[IdTree \%in\% "100658"]

# Inputs
Time <-  c(2000, 2002, 2004, 2006, 2008, 2010)
DBHCor <- c(13, 14, 15, 12, 13, 14)
cresc <- c(0.5, 0.5, NA, 0.5, 0.5)
cresc_abs <- c(1, 1, NA, 1, 1)
cresc_abn <- 3

CorrectionType = "phylogenetic hierarchical"

DataTree_output <- PhylogeneticHierarchicalCorrection(DataTree = DataTree,
                                                      Data = TestData,
                                                      cresc = cresc, cresc_abs = cresc_abs, cresc_abn = cresc_abn,
                                                      DBHCor = DBHCor, Time = Time,
                                                      PositiveGrowthThreshold = 5,
                                                      NegativeGrowthThreshold = -2,
                                                      DBHRange = 10, MinIndividualNbr = 5
)
}
