% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PunctualErrorDetection.R
\name{PunctualErrorDetection}
\alias{PunctualErrorDetection}
\title{PunctualErrorDetection}
\usage{
PunctualErrorDetection(
  DBHCor,
  Time,
  PositiveGrowthThreshold = 5,
  NegativeGrowthThreshold = -2,
  DetectOnly = FALSE
)
}
\arguments{
\item{DBHCor}{Diameter vector (numeric)}

\item{Time}{Time vector (numeric)}

\item{PositiveGrowthThreshold}{in cm/year : a tree
widening by more than x cm/year is considered abnormal (numeric, 1 value)}

\item{NegativeGrowthThreshold}{in cm/census : The possible
positive measurement error (+n) cannot be corrected until the growth
appears abnormal, but a negative measurement error can be allowed until -n
(a tree does not decrease). Thus the positive measurement error (+n) is
"compensated". (numeric, 1 value)}

\item{DetectOnly}{TRUE: Only detect errors, FALSE: detect and correct errors
(Default: FALSE) (logical)}
}
\value{
The diameter vector (numeric) with NA instead of the punctual errors
detected. Return the intact input diameter vector in case of shift errors.
}
\description{
PunctualErrorDetection
}
\examples{
Time <- c(2000, 2002, 2004, 2006, 2008, 2012, 2014, 2016, 2020)

# 1 ind at a time
## Punctual error case
DBHCor <- c(13, 14, 15, 16, 30, 19, 15, 21, 23)
plot(Time, DBHCor)

PunctualErrorDetection(DBHCor = DBHCor, Time = Time)

## 2 shifts error case
DBHCor <- c(13, 14, 15, 16, 12, 14, 15, 11, 13)
plot(Time, DBHCor)

PunctualErrorDetection(DBHCor = DBHCor, Time = Time)

}
