% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecruitmentCorrection.R
\name{RecruitmentCorrectionByTree}
\alias{RecruitmentCorrectionByTree}
\title{RecruitmentCorrectionByTree}
\usage{
RecruitmentCorrectionByTree(
  DataTree,
  MinDBH = 10,
  PositiveGrowthThreshold = 5,
  InvariantColumns = c("Site", "Genus", "Species", "Family", "ScientificName"),
  PlotCensuses,
  DetectOnly = FALSE
)
}
\arguments{
\item{DataTree}{A dataset corresponding to a single tree's (1 IdTree)
measurements (data.frame or data.table)}

\item{MinDBH}{Minimum diameter of trees inventoried according to your
protocol (in cm) (numeric, 1 value)}

\item{PositiveGrowthThreshold}{A tree widening by more than x
cm/year is considered abnormal (numeric, 1 value)}

\item{InvariantColumns}{Vector with the names of the columns that are
supposed to have always the same value for each measurement of the same
tree (character)}

\item{PlotCensuses}{Census years for the plot in which the tree is (integer)}

\item{DetectOnly}{TRUE: Only detect errors, FALSE: detect and correct errors
(logical)}
}
\value{

}
\description{
RecruitmentCorrectionByTree
}
\examples{
library(data.table)
DataTree <- data.table(IdTree = "a",
                       CensusYear = seq(2000,2008, by = 2), # 2 years/census
                       DBHCor  = as.numeric(c(13:17)), # 1cm/census(0.5 cm/year)
                       Site = "Imaginary forest"
)
# 1st DBH = 13 > MinDBH. In 1998 the tree was 12cm, 11cm in 1996 and 10cm in
1994.
RecruitmentCorrectionByTree(DataTree,
                            InvariantColumns = "Site",
                            PlotCensuses = seq(1996,2016, by = 2))

}
