% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RequireFormat_interactive.R
\name{RequiredFormat_interactive}
\alias{RequiredFormat_interactive}
\title{RequiredFormat_interactive}
\usage{
RequiredFormat_interactive(Data, input = list())
}
\arguments{
\item{Data}{Input forest inventory (data.frame)}

\item{input}{a named list. Usually empty, except in Shiny App}
}
\value{
a names list with all the argument needed for RequiredFormat_calcul function (column name correspondances + extra input)
}
\description{
RequiredFormat_interactive
}
\examples{
\dontrun{
data(ParacouSubset)
ParacouProfile <- RequiredFormat_interactive(ParacouSubset)
               }

}
