% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReversedRequiredFormat.R
\name{ReversedRequiredFormat}
\alias{ReversedRequiredFormat}
\title{ReversedRequiredFormat}
\usage{
ReversedRequiredFormat(Data, input, x = NULL, ThisIsShinyApp = FALSE)
}
\arguments{
\item{Data}{Standardized data, returned from \link{RequiredFormat}}

\item{input}{A named list, typically the output of function
RequiredFormat_interactive, also called site profile. It has information on
column names correspondence, size units etc...}

\item{x}{For internal use when function used by Shiny app}

\item{ThisIsShinyApp}{For internal use when function used by Shiny app
(logical)}
}
\value{
(data.frame) in the format given the profile selected in input.
}
\description{
ReversedRequiredFormat
}
\details{
This function takes the standardized forest inventory data.table (returned by \link{RequiredFormat})
and converts the column names to the names of the profile given as \link{input}
}
\examples{
\dontrun{

data(ParacouSubsetFormated)
data("ForestGEOProfile")
ReversedRequiredFormat(ParacouSubsetFormated, ForestGEOProfile)
               }

}
