% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunApp.R
\name{RunApp}
\alias{RunApp}
\title{RunApp}
\usage{
RunApp()
}
\value{
Opens an interactive ShinnyApp window
}
\description{
RunApp
}
\details{
This function launches the application located in inst/ directory
}
\examples{
\dontrun{
RunApp()
               }
}
