% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Select2FramingValues.R
\name{Select2FramingValues}
\alias{Select2FramingValues}
\title{Select the 2 framing values around a given value}
\usage{
Select2FramingValues(i, Y, X)
}
\arguments{
\item{i}{Position of the value from which to find the framing values}

\item{Y}{Vector of values of the variable from which to select the 2 values
around i (numeric)}

\item{X}{Vector of values of the explanatory variable of the variable Y (thus
also to filter) (numeric)}
}
\value{
2 vectors of 2 values each (list):
\itemize{
\item yval: the 2 Y values around i
\item xval: the 2 X values corresponding to the two yval values
}
}
\description{
Select the 2 framing values around a given value
}
\examples{
DBHCor = c(17.7, 26.2, NA, 34.6, 34.6, 34.6, 38.0)
Time = c(1998, 2008, 2016, 2017, 2018, 2019, 2021)

Select2FramingValues(i = 3, Y = DBHCor, X = Time)

}
