% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeData-package.R
\docType{package}
\name{TreeData}
\alias{TreeData}
\alias{TreeData-package}
\title{TreeData-package}
\description{
Forest Inventories Harmonization & Correction
}
\section{TreeData functions}{

RequiredFormat
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/VincyaneBadouard/TreeData.git}
  \item Report bugs at \url{https://github.com/VincyaneBadouard/TreeData/issues}
}

}
\author{
\strong{Maintainer}: Vincyane Badouard \email{vincyane.badouard@gmail.com}

Authors:
\itemize{
  \item Valentine Herrmann \email{HerrmannV@si.edu}
  \item Camille Piponiot \email{camille.piponiot@gmail.com}
}

Other contributors:
\itemize{
  \item Nino Page \email{page.nino@gmail.com} [contributor]
}

}
\keyword{internal}
