% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoValDiameterCor.R
\name{TwoValDiameterCor}
\alias{TwoValDiameterCor}
\title{Correction of 2 diameter values}
\usage{
TwoValDiameterCor(
  DataTree,
  Data,
  DBHCor,
  Time,
  CorrectionType = NULL,
  PositiveGrowthThreshold = 5,
  NegativeGrowthThreshold = -2,
  DBHRange,
  MinIndividualNbr,
  OtherCrit,
  DetectOnly = FALSE
)
}
\arguments{
\item{DataTree}{A dataset corresponding to a single tree/stem's (1
IdTree/IdStem) measurements (data.table)
If 'phylogenetic hierarchical' correction is asked, the dataset must
contain the columns:
\itemize{
\item \code{IdStem} (character)
\item \code{ScientificNameCor} (character)
\item \code{GenusCor} (character)
\item \code{FamilyCor} (character)
\item \code{Diameter} (numeric)
\item \code{Year} (numeric)
}}

\item{Data}{Complete dataset (data.table)
The dataset must contain the columns:
\itemize{
\item \code{IdStem} (character)
\item \code{ScientificNameCor} (character)
\item \code{GenusCor} (character)
\item \code{FamilyCor} (character)
\item \code{Diameter} (numeric)
\item \code{Year} (numeric)
}}

\item{DBHCor}{Diameter vector in cm (numeric)}

\item{Time}{Time vector in years (numeric)}

\item{CorrectionType}{Possible values:
\itemize{
\item NULL: The 1st non-NA diameter value is assigned to the 2nd non-NA diameter
value
\item "phylogenetic hierarchical": replace abnormal growth with the average
growth of other trees in the dataset, at the specific, genus, family or
stand level, within a DBH range of x cm (\emph{DBHRange} argument). If the
number of these trees < n (\emph{MinIndividualNbr} argument) at the specific
level, we switch to the genus level etc.
}}

\item{PositiveGrowthThreshold}{in cm/year : a tree
widening by more than x cm/year is considered abnormal (numeric, 1 value)}

\item{NegativeGrowthThreshold}{in cm/census : The possible
positive measurement error (+n) cannot be corrected until the growth
appears abnormal, but a negative measurement error can be allowed until -n
(a tree does not decrease). Thus the positive measurement error (+n) is
"compensated". (numeric, 1 value)}

\item{DBHRange}{DBH range in cm to take into account to select other trees in
the dataset to apply "phylogenetic hierarchical" correction (Default: 10
cm) (numeric, 1 value)}

\item{MinIndividualNbr}{Minimum number of individuals to take into account in
"phylogenetic hierarchical" correction (Default: 5) (numeric, 1 value)}

\item{OtherCrit}{Other criteria to select the individuals used for the
calculation of the mean growth. Give the name of the column(s) for which the
individuals must have the same value as the tree to correct (e.g. c("Plot",
"Subplot")) (character)}

\item{DetectOnly}{TRUE: Only detect errors, FALSE: detect and correct errors
(Default: FALSE) (logical)}
}
\value{
List of 2 objects:
\itemize{
\item DBHCor (numeric vector): corrected by the method
\item DataTree (data.table): with the \emph{DiameterCorrectionMeth} column filled with
correction method.
}
}
\description{
Correction of 2 diameter values
}
\examples{
library(data.table)
data(TestData)

DataTree <- TestData[IdTree \%in\% "100771"]
DataTree <- DataTree[order(Year)] # order de dt

DataTree$Diameter <- c(NA, 13, NA, 8, NA)

DBHCor <- DataTree$Diameter
Time <- DataTree$Year

TwoValDiameterCor(DataTree,
                  Data = TestData,
                  DBHCor, Time,
                  CorrectionType = "phylogenetic hierarchical",
                  DBHRange = 10,
                  MinIndividualNbr = 1, OtherCrit = NULL)

}
