% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WFOdataDubset.R
\docType{data}
\name{WFOdataSubset}
\alias{WFOdataSubset}
\title{WFOdataSubset}
\format{
A data.table with 351 rows and 19 variables:
\describe{
\item{taxonID}{(character)}
\item{scientificNameID}{(character)}
\item{scientificName}{(character)}
\item{taxonRank}{(character)}
\item{parentNameUsageID}{(character)}
\item{scientificNameAuthorship}{(character)}
\item{family}{(character)}
\item{genus}{(character)}
\item{specificEpithet}{(character)}
\item{infraspecificEpithet}{(character)}
\item{verbatimTaxonRank}{(character)}
\item{nomenclaturalStatus}{(character)}
\item{namePublishedIn}{(character)}
\item{taxonomicStatus}{(character)}
\item{acceptedNameUsageID}{(character)}
\item{nameAccordingToID}{(character)}
\item{created}{(Date)}
\item{modified}{(Date)}
\item{references}{(character)}
...
}
}
\source{
\url{http://www.worldfloraonline.org/downloadData}
}
\usage{
WFOdataSubset
}
\description{
This data set is a subset of the World Flora Online taxonomic backbone that
allows running the TreeData package tests. In practical applications, users
should first download a static copy of the Taxonomic Backbone data from
http://www.worldfloraonline.org/downloadData .
}
\keyword{datasets}
