% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BotanicalCorrection.R
\name{BotanicalCorrection}
\alias{BotanicalCorrection}
\title{Botanical Correction}
\usage{
BotanicalCorrection(Data, DetectOnly = FALSE)
}
\arguments{
\item{Data}{Dataset (data.frame or data.table)
The dataset must contain the columns:
\itemize{
\item \code{IdTree} (character)
\item \code{Family} (character)
\item \code{Genus} (character)
\item \code{Species} (character)
\item \code{VernName} (character)
\item \code{ScientificName} (character)
}}

\item{DetectOnly}{TRUE: Only detect errors, FALSE: detect and correct errors
(Default: FALSE) (logical)}
}
\value{
Fill the \emph{Comment} column with error type informations. If
\emph{DetectOnly} = FALSE, add columns:
\itemize{
\item \code{FamilyCor} (character): corrected Family name
\item \code{FamilyCorSource} (character): source of the Family correction
\item \code{GenusCor} (character): corrected Genus name
\item \code{SpeciesCor} (character): corrected Species name
\item \code{BotanicalCorrectionSource} (character): source of the Genus and Species
correction
\item \code{ScientificNameCor} (character): corrected Scientific name
\item \code{VernNameCor} (character): completed if information available at \code{IdTree}
level.
}
}
\description{
Botanical Correction
}
\examples{
library(data.table)
data(TestData)

Rslt <- BotanicalCorrection(TestData)

ScfcCor <- unique(Rslt[ScientificNameCor != ScientificName,
                list(ScientificName, ScientificNameCor,
                Family, FamilyCor, FamilyCorSource,
                Genus, GenusCor,
                Species, SpeciesCor,
                BotanicalCorrectionSource, Comment)
                ])

FamCor <- unique(Rslt[FamilyCor != Family,
                list(ScientificName, ScientificNameCor,
                Family, FamilyCor, FamilyCorSource,
                Genus, GenusCor,
                Species, SpeciesCor,
                BotanicalCorrectionSource, Comment)
                ])

}
