% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateComment.R
\name{CommentByRow}
\alias{CommentByRow}
\title{CommentByRow}
\usage{
CommentByRow(row, comment)
}
\arguments{
\item{row}{A 1 row data.table (data.table)}

\item{comment}{The string to add in the "Comment" column (character)}
}
\value{
The input data.table with the filled "Comment" column.
}
\description{
Add a string in the "Comment" column of a data.table.
If there is already a value, the string is pasted after the "/" separator.
}
\examples{
\dontrun{
library(data.table)
dt <- data.table(A = c(2), B = c(6))

dt <- CommentByRow(dt, comment = "A = 2") # 1st comment

CommentByRow(dt, comment = "B = 6")
}
}
