% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeIncrementation.R
\name{ComputeIncrementation}
\alias{ComputeIncrementation}
\title{ComputeIncrementation}
\usage{
ComputeIncrementation(Var, Type = "annual", Time)
}
\arguments{
\item{Var}{Variable (numeric)}

\item{Type}{Type of incrementation: "absolute" or "annual" (character)
\itemize{
\item "absolute": absolute increment (not divided by time between 2 values)
\item "annual": annual increment
}}

\item{Time}{Time variable (numeric)}
}
\value{
Absolute or annual incrementation of the variable
}
\description{
ComputeIncrementation
}
\examples{

DBHCor = c(13, 14, 15, 16, 30, 19, 15, 21, 23)
Time = c(2000, 2002, 2004, 2006, 2008, 2012, 2014, 2016, 2020)
Rslt <- ComputeIncrementation(Var = DBHCor, Time = Time)

}
