% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiameterCorrection.R
\name{DiameterCorrection}
\alias{DiameterCorrection}
\title{Diameter correction}
\usage{
DiameterCorrection(
  Data,
  DefaultHOM = 1.3,
  MinDBH = 10,
  MaxDBH = 500,
  PositiveGrowthThreshold = 5,
  NegativeGrowthThreshold = -2,
  Pioneers = c("Cecropia", "Pourouma"),
  PioneersGrowthThreshold = 7.5,
  TrustMeasSet = "first",
  WhatToCorrect = c("POM change", "punctual", "shift"),
  CorrectionType = c("taper", "quadratic", "linear", "individual",
    "phylogenetic hierarchical"),
  DBHRange = 10,
  MinIndividualNbr = 5,
  Digits = 1L,
  TaperParameter = function(DAB, HOM) 0.156 - 0.023 * log(DAB) - 0.021 * log(HOM),
  TaperFormula = function(DAB, HOM, TaperParameter) DAB/(2 * exp(-TaperParameter * (HOM
    - DefaultHOM))),
  DetectOnly = FALSE
)
}
\arguments{
\item{Data}{(data.frame or data.table)
The dataset must contain the columns:
\itemize{
\item 'IdTree' (character)
\item 'Diameter' (numeric)
\item 'Year' (numeric)
\item \strong{'HOM'(Height Of Measurement) (numeric)} if you want to apply the
\strong{"taper" correction}
\item \strong{'POM'(Point Of Measurement) (factor)} if you want to correct from the
\strong{"POM change}
If you want to apply the \strong{"phylogenetic hierarchical"} correction, the
dataset must contain the columns:
\item 'ScientificName' (character)
\item 'Genus' (character)
\item 'Family' (character)
}}

\item{DefaultHOM}{Default Height Of Measurement in meter (Default: 1.3 m)
(numeric, 1 value)}

\item{MinDBH}{Minimum census DBH of your protocol in cm (numeric, 1 value)}

\item{MaxDBH}{Maximum possible DBH of your stand in cm (numeric, 1 value)}

\item{PositiveGrowthThreshold}{in cm/year : a tree
widening by more than x cm/year is considered abnormal (numeric, 1 value)}

\item{NegativeGrowthThreshold}{in cm/census : The possible
positive measurement error (+n) cannot be corrected until the growth
appears abnormal, but a negative measurement error can be allowed until -n
(a tree does not decrease). Thus the positive measurement error (+n) is
"compensated". (numeric, 1 value)}

\item{Pioneers}{(characters vector)}

\item{PioneersGrowthThreshold}{in cm (numeric, 1 value)}

\item{TrustMeasSet}{Trust measurements set: the "first" or the "last" set
(character, 1 value)}

\item{WhatToCorrect}{c("POM change", "punctual", "shift") (character)
\itemize{
\item "POM change": detect POM change in the column 'POM' or 'HOM' and correct
the Diameter values from it.
\item "punctual": detect if the error is punctual and correct it by
interpolation.
\item "shift": detect if there is a shift of several Diameter values and links them
to the trust measurements set (\emph{TrustMeasSet} argument).
}}

\item{CorrectionType}{c("taper", "linear", "quadratic", "individual",
"phylogenetic hierarchical") (character).
\itemize{
\item "taper": correct for biases associated with nonstandard and changing
measurement heights, from a taper model (\emph{TaperParameter} &
\emph{TaperFormula} arguments).
Correction possible only if the 'HOM' (Height Of Measurement)
column is available.
\item "linear": interpolation by linear regression of the individual annual
growth over time.
\item "quadratic": interpolation by quadratic regression  of the individual
annual growth over time.
\item "individual": replace abnormal growth by interpolation from the
individual values.
\item "phylogenetic hierarchical": replace abnormal growth with the average
growth of other trees in the dataset, at the specific, genus, family
or stand level, within a DBH range of x cm (\emph{DBHRange} argument).
If the number of these trees < n (\emph{MinIndividualNbr} argument)
at the specific level, we switch to the genus level etc.
}}

\item{DBHRange}{DBH range in cm to take into account to select other trees in
the dataset to apply "phylogenetic hierarchical" correction (Default: 10
cm) (numeric, 1 value)}

\item{MinIndividualNbr}{Minimum number of individuals to take into account in
"phylogenetic hierarchical" correction (Default: 5) (numeric, 1 value)}

\item{Digits}{Number of decimal places to be used in the 'DBHCor' column
(Default: 1L) (integer)}

\item{TaperParameter}{Taper parameter (unitless) formula (function)
Default: \emph{TaperParameter = 0.156 - 0.023 log(DAB) - 0.021 log(HOM)}
of Cushman et al.2021.
With:
\itemize{
\item \emph{DAB}: Diameter Above Buttress (in cm)
\item \emph{HOM}: Height Of Measurement (in m)
}}

\item{TaperFormula}{Taper formula (function)
Default: \emph{DAB / (2 e^(- TaperParameter (HOM - DefaultHOM)))}
of Cushman et al.2021.
With:
\itemize{
\item \emph{DAB}: Diameter Above Buttress (in cm)
\item \emph{HOM}: Height Of Measurement (in m)
\item \emph{DefaultHOM}:  Default Height Of Measurement (in m)
\item \emph{TaperParameter}: Taper parameter (unitless)
}}

\item{DetectOnly}{TRUE: Only detect errors, FALSE: detect and correct errors
(Default: FALSE) (logical)}
}
\value{
Fill the \emph{Comment} column with error type informations. If
\emph{DetectOnly} = FALSE, add columns: - \emph{DBHCor}: corrected trees diameter at
default HOM - \emph{DiameterCorrectionMeth} =
"taper"/"linear"/"quadratic"/"individual"/"phylogenetic hierarchical"
}
\description{
Diameter correction
}
\examples{
library(data.table)
data(TestData)

# Rslt <- DiameterCorrection(
#  TestData,
#  CorrectionType = c("quadratic",
#                     "individual", "phylogenetic hierarchical"))

}
