% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiameterCorrection.R
\name{DiameterCorrection}
\alias{DiameterCorrection}
\title{Diameter correction}
\usage{
DiameterCorrection(
  Data,
  UseTaperCorrection = TRUE,
  DefaultHOM = 1.3,
  TaperParameter = function(DAB, HOM) 0.156 - 0.023 * log(DAB) - 0.021 * log(HOM),
  TaperFormula = function(DAB, HOM, TaperParameter, DefaultHOM) DAB/(exp(-TaperParameter
    * (HOM - DefaultHOM))),
  KeepMeas = c("MaxHOM", "MaxDate"),
  PositiveGrowthThreshold = 5,
  NegativeGrowthThreshold = -2,
  Pioneers = NULL,
  PioneersGrowthThreshold = 7.5,
  WhatToCorrect = c("POM change", "Abnormal growth"),
  CorrectionType = c("individual", "phylogenetic hierarchical"),
  DBHRange = 10,
  MinIndividualNbr = 5,
  OtherCrit = NULL,
  DBHCorForDeadTrees = TRUE,
  coef = 0.9
)
}
\arguments{
\item{Data}{Dataset (data.frame or data.table)
The dataset must contain the columns:
\itemize{
\item \code{IdTree} or \code{IdStem} (character)
\item \code{ScientificName_TreeDataCor} (character)
\item \code{Diameter} (numeric)
\item \code{IdCensus} (ordered factor)
\item \strong{\code{POM} (Point Of Measurement) (factor)} or
\strong{\code{HOM} (Height Of Measurement) (numeric)} if you want to correct from
the \strong{"POM change"}
If you want to apply the \strong{"phylogenetic hierarchical"} correction, the
dataset must also contain the columns:
\item \code{Genus_TreeDataCor} (character)
\item \code{Family_TreeDataCor} (character)
}}

\item{UseTaperCorrection}{(logical) TRUE: transform the tree diameter measured at a given height
into the diameter corresponding to the default measurement height (\code{DefaultHOM}), using an allometry.
FALSE: do not apply a taper correction}

\item{DefaultHOM}{Default Height Of Measurement in meter (Default: 1.3 m)
(numeric, 1 value)}

\item{TaperParameter}{Taper parameter (unitless) formula (function)
Default: \emph{TaperParameter = 0.156 - 0.023 log(DAB) - 0.021 log(HOM)}
of Cushman et al.2021.
With:
\itemize{
\item \emph{DAB}: Diameter Above Buttress (in cm)
\item \emph{HOM}: Height Of Measurement (in m)
}}

\item{TaperFormula}{Taper formula (function)
Default: \emph{DAB / (e^(- TaperParameter (HOM - DefaultHOM)))}
of Cushman et al.2021.
With:
\itemize{
\item \emph{DAB}: Diameter Above Buttress (in cm)
\item \emph{HOM}: Height Of Measurement (in m)
\item \emph{DefaultHOM}:  Default Height Of Measurement (in m)
\item \emph{TaperParameter}: Taper parameter (unitless)
}}

\item{KeepMeas}{In case of \strong{multiple diameter measurements} in the same
census:
Possible values: "MaxHOM", "MaxDate" (character).
\itemize{
\item "MaxHOM": apply the correction to the measurement taken at the
\strong{highest POM}
\item "MaxDate": apply the correction to the \strong{most recent measurement} (same
IdCensus but more recent date)
}}

\item{PositiveGrowthThreshold}{in cm/year: a tree
widening by more than this value is considered abnormal (numeric, 1 value)}

\item{NegativeGrowthThreshold}{in cm/census: the possible
positive measurement error (+n) cannot be corrected until the growth
appears abnormal, but a negative measurement error can be allowed until -n
(a tree does not decrease). Thus the positive measurement error (+n) is
"compensated". (numeric, 1 value)}

\item{Pioneers}{Scientific names of the pioneer species of the site, as in
the \code{ScientificName_TreeDataCor} column (characters vector)}

\item{PioneersGrowthThreshold}{in cm/year: a tree of a pioneer species that
widens by more than this value is considered abnormal (numeric, 1 value)}

\item{WhatToCorrect}{Possible values: "POM change", "Abnormal growth"
(character). All are complementary and recommended.
\itemize{
\item "POM change": detect POM change in the column \code{POM} and correct the
Diameter values from it. (Ignored if taper correction is applied)
\item "Abnormal growth": detect if the growth is greater than PositiveGrowthThreshold ('PioneersGrowthThreshold' if species belongs to 'Pioneers')
or smaller than NegativeGrowthThreshold and correct it by \code{CorrectionType}
}}

\item{CorrectionType}{Possible values: "individual", "phylogenetic
hierarchical" (character, 1 value).
\itemize{
\item "individual": replace abnormal growth by interpolation from the
individual values.
\item "phylogenetic hierarchical": replace abnormal growth with the average
growth of other trees in the dataset, at the specific, genus, family
or stand level, within a DBH range of x cm (\emph{DBHRange} argument).
If the number of these trees < n (\emph{MinIndividualNbr} argument)
at the specific level, we switch to the genus level etc.
}}

\item{DBHRange}{DBH range in cm to take into account to select other trees in
the dataset to apply "phylogenetic hierarchical" correction (Default: 10
cm) (numeric, 1 value)}

\item{MinIndividualNbr}{Minimum number of individuals to take into account in
"phylogenetic hierarchical" correction (Default: 5) (numeric, 1 value)}

\item{OtherCrit}{Other criteria to select the individuals used for the
calculation of the mean growth in the "phylogenetic hierarchical"
correction. Give the name of the column(s) for which the individuals must
have the same value as the tree to correct (e.g. c("Plot", "Subplot"))
(character)}

\item{DBHCorForDeadTrees}{(logical) TRUE: return DBHCor also for dead trees.
FALSE: do not return DBHCor for dead trees. In this case it is advisable to
have corrected the tree life status with the \emph{StatusCorrection()} function.}

\item{coef}{(numeric, 1 value) This is used in individual corrections, to calculate weight of the growths by temporal proximity}
}
\value{
Fill the \emph{Comment_TreeData} column with error type informations and add columns:
\itemize{
\item \emph{Diameter_TreeDataCor}: corrected trees diameter at default HOM
\item \emph{DiameterCorrectionMeth_TreeData} = "local linear regression","weighted
mean"/phylogenetic hierarchical("species"/"genus"/"family"/"stand")/
"shift realignment"/"Same value".
\item \emph{POM_TreeDataCor} (factor): POM value at which the corrected diameters are proposed.
Corresponds to the 1st POM value at which the stem was measured.
\item \emph{HOM_TreeDataCor} (numeric): HOM value at which the corrected diameters
are proposed. Corresponds to the 1st HOM value at which the stem was
measured.
}
}
\description{
Diameter correction
}
\details{
When there is only 1 \code{Diameter} value for a tree/stem,
\code{Diameter_TreeDataCor} takes the original \code{Diameter} value. If this value
is 0 or > MaxDBH, \code{Diameter_TreeDataCor} takes NA. Diameters not linked to
an IdTree/IdStem or to a Census IdCensus are not processed.
Punctual error correction only with linear regression and not quadratic,
because punctual errors are corrected from a local regression with the 2
framing values.
}
\examples{
# library(data.table)
data(TestData)

TestData$HOM[1:3] <- c(0.5,1.5,NA)
TestData$Diameter[21:23] <- c(31,91,14)

Rslt <- DiameterCorrection(
 TestData,
  WhatToCorrect = c("POM change", "Abnormal growth"),
    CorrectionType = c("phylo"),
    MinIndividualNbr = 1)

DiameterCorrectionPlot(Rslt, OnlyCorrected = TRUE)

}
