% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiameterCorrectionPlot.R
\name{DiameterCorrectionPlot}
\alias{DiameterCorrectionPlot}
\title{Plot diameter correction result}
\usage{
DiameterCorrectionPlot(Data, OnlyCorrected = FALSE, CorCol = "DBHCor")
}
\arguments{
\item{Data}{Dataset (data.frame or data.table)
The dataset must contain the columns:
\itemize{
\item \code{IdStem} (character)
\item \code{Year} (numeric)
\item \code{Diameter} (numeric)
\item \code{DBHCor} (numeric)
\item \code{HOM} (Height Of Measurement) (numeric)
\item \code{HOMCor} (Corrected Height Of Measurement) (numeric)
}}

\item{OnlyCorrected}{TRUE: plot only corrected stems, FALSE: plot all stems}

\item{CorCol}{Corrected column name (character)}
}
\value{
The plots of the initial diameter values and proposed corrections, by
IdStem.
}
\description{
Plot diameter correction result
}
\examples{

\dontrun{
pdf("DiameterCorrectionPlots_TestData.pdf", width = 25, height = 10)
DiameterCorrectionPlot(Rslt, OnlyCorrected = TRUE)
dev.off()
}

}
