% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErrorsDetection.R
\name{ErrorsDetection}
\alias{ErrorsDetection}
\title{Errors Detection}
\usage{
ErrorsDetection(
  Data,
  ByStem = FALSE,
  DeathConfirmation = 2,
  UseSize = FALSE,
  MinDBH = 10,
  PositiveGrowthThreshold = 5
)
}
\arguments{
\item{Data}{Dataset (data.frame or data.table)}

\item{ByStem}{must be equal to TRUE if your inventory contains the stem
level, equal to FALSE if not, and in this case the correction is done by
tree (logical)}

\item{DeathConfirmation}{Number of times (censuses) needed for an unseen tree
to be considered dead (numeric) (Default = 2 censuses)}

\item{UseSize}{Use the size presence as a witness of the living status of the
tree (logical) (Default = FALSE)}

\item{MinDBH}{Minimum diameter of trees inventoried according to your
protocol (in cm) (numeric, 1 value) (Default = 10 cm)}

\item{PositiveGrowthThreshold}{in cm/year: a tree
widening by more than this value is considered abnormal (numeric, 1 value)}
}
\value{
The input dataset (data.table) with a new \emph{Comment} column with error
type informations.
}
\description{
Errors Detection
}
\details{
Detect errors
\itemize{
\item Remove \strong{duplicated rows}
\item Check \strong{missing value} in
X-Yutm/PlotArea/Plot/SubPlot/Year/TreeFieldNum/
IdTree/IdStem/Diameter/POM/HOM/Family/Genus/Species/VernName
\item Check \strong{missing value} (NA/0) in the measurement variables: "Diameter",
"HOM", "TreeHeight", "StemHeight"
\item Check \strong{duplicated TreeFieldNum} in plot-subplot association in a census
(at the site scale)
\item Check of the \emph{unique association of the idTree with plot, subplot and
TreeFieldNum} (at the site scale)
\item Check \strong{duplicated idTree/IdStem} in a census (at the site scale)
\item Check for trees \strong{outside the subplot}
\item Check \strong{invariant coordinates per IdTree/IdStem}
\item Check \strong{fix Plot and SubPlot number}
\item Check botanical identification (\emph{BotanicalCorrection})
\item Check the life status evolution of the trees (\emph{StatusCorrection})
\item Check diameter evolution of the trees (\emph{DiameterCorrection})
\item Check recruitment (\emph{RecruitmentCorrection})
}
}
\examples{
library(data.table)
data("TestData")

Rslt <- ErrorsDetection(TestData)

}
\seealso{
\code{\link{StatusCorrection}}
}
