% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForestGEOProfile.R
\docType{data}
\name{ForestGeoProfile}
\alias{ForestGeoProfile}
\title{ForestGeoProfile}
\format{
A named list
\describe{
\item{TreeHeightUnitMan}{Value or column name in data set @ForestGeoSubset (none) corresponding to What is your tree height unit?}
\item{Family}{Value or column name in data set @ForestGeoSubset (Family) corresponding to Family name}
\item{TreeHeight}{Value or column name in data set @ForestGeoSubset (none) corresponding to Tree height}
\item{Date}{Value or column name in data set @ForestGeoSubset (ExactDate) corresponding to Date of measurement}
\item{Month}{Value or column name in data set @ForestGeoSubset (none) corresponding to Month of measurement}
\item{Day}{Value or column name in data set @ForestGeoSubset (none) corresponding to Day of measurement}
\item{POM}{Value or column name in data set @ForestGeoSubset (none) corresponding to POM}
\item{HOM}{Value or column name in data set @ForestGeoSubset (hom) corresponding to HOM}
\item{PlotAreaUnitMan}{Value or column name in data set @ForestGeoSubset (none) corresponding to What is the unit for plot area?}
\item{Xutm}{Value or column name in data set @ForestGeoSubset (none) corresponding to Tree X UTM coordinate}
\item{IdStem}{Value or column name in data set @ForestGeoSubset (stemID) corresponding to Stem ID}
\item{PlotMan}{Value or column name in data set @ForestGeoSubset (ForestGEO) corresponding to What is your plot name ?}
\item{Species}{Value or column name in data set @ForestGeoSubset (Species) corresponding to Species name}
\item{ScientificNameSep}{Value or column name in data set @ForestGeoSubset () corresponding to Character separating Genus and Species in Scientific name}
\item{Yplot}{Value or column name in data set @ForestGeoSubset (gy) corresponding to Tree Y euclidean position in plot}
\item{Ysubplot}{Value or column name in data set @ForestGeoSubset (none) corresponding to Tree Y euclidean position in SUBplot}
\item{CircUnitMan}{Value or column name in data set @ForestGeoSubset (none) corresponding to What is your circumference unit?}
\item{BD}{Value or column name in data set @ForestGeoSubset (none) corresponding to BD}
\item{BDUnitMan}{Value or column name in data set @ForestGeoSubset (none) corresponding to What is your BD unit ?}
\item{BPOM}{Value or column name in data set @ForestGeoSubset (none) corresponding to Basal POM}
\item{BHOM}{Value or column name in data set @ForestGeoSubset (none) corresponding to Basal HOM}
\item{BHOMUnitMan}{Value or column name in data set @ForestGeoSubset (none) corresponding to What is your BasalHOM unit?}
\item{BCirc}{Value or column name in data set @ForestGeoSubset (none) corresponding to Basal Circumference}
\item{BCircUnitMan}{Value or column name in data set @ForestGeoSubset (none) corresponding to What is your basal circumference unit?}
\item{IdMeasure}{Value or column name in data set @ForestGeoSubset (MeasureID) corresponding to Measure ID}
\item{IdLevel}{Value or column name in data set @ForestGeoSubset (IDLevel) corresponding to Deepest taxonomic level for which full identification is known}
\item{TreeFieldNum}{Value or column name in data set @ForestGeoSubset (tag) corresponding to Tree tag}
\item{ScientificName}{Value or column name in data set @ForestGeoSubset (Latin) corresponding to Scientific name}
\item{DateFormat}{Value or column name in data set @ForestGeoSubset (yyyy-mm-dd) corresponding to What is your date format?}
\item{IdCensus}{Value or column name in data set @ForestGeoSubset (CensusID) corresponding to Census ID}
\item{PlotAreaMan}{Value or column name in data set @ForestGeoSubset (25.6) corresponding to What is your plot area in hectare?}
\item{Year}{Value or column name in data set @ForestGeoSubset (none) corresponding to Year of measurement}
\item{LifeStatus}{Value or column name in data set @ForestGeoSubset (DFstatus) corresponding to Life status}
\item{PlotArea}{Value or column name in data set @ForestGeoSubset (none) corresponding to Plot area}
\item{CommercialSp}{Value or column name in data set @ForestGeoSubset (none) corresponding to Commercial species indicator}
\item{SubPlotAreaUnitMan}{Value or column name in data set @ForestGeoSubset (none) corresponding to What is the unit for subplot area?}
\item{Yutm}{Value or column name in data set @ForestGeoSubset (none) corresponding to Tree Y UTM coordinate}
\item{utmUnitMan}{Value or column name in data set @ForestGeoSubset (none) corresponding to UTM units}
\item{YearMan}{Value or column name in data set @ForestGeoSubset (NA) corresponding to What is your year of measurement?}
\item{Circ}{Value or column name in data set @ForestGeoSubset (none) corresponding to Circumference}
\item{SubPlotAreaMan}{Value or column name in data set @ForestGeoSubset (0.04) corresponding to What is your subplot area in hectare?}
\item{SubPlotMan}{Value or column name in data set @ForestGeoSubset () corresponding to What is your subplot name ?}
\item{Site}{Value or column name in data set @ForestGeoSubset (none) corresponding to Site name}
\item{Lon}{Value or column name in data set @ForestGeoSubset (none) corresponding to Tree longitude}
\item{Plot}{Value or column name in data set @ForestGeoSubset (none) corresponding to Plot name}
\item{IdTree}{Value or column name in data set @ForestGeoSubset (treeID) corresponding to Tree ID}
\item{LifeForm}{Value or column name in data set @ForestGeoSubset (none) corresponding to Life form}
\item{VernName}{Value or column name in data set @ForestGeoSubset (none) corresponding to Vernacular name}
\item{SubPlotArea}{Value or column name in data set @ForestGeoSubset (none) corresponding to SubPlot area}
\item{SiteMan}{Value or column name in data set @ForestGeoSubset (SCBI) corresponding to What is your site name ?}
\item{IsLive}{Value or column name in data set @ForestGeoSubset (c("alive", "broken below")) corresponding to Which of your status(es) represent a LIVE tree?}
\item{Authority}{Value or column name in data set @ForestGeoSubset (SpeciesID) corresponding to  Taxonomic authority for the ID level}
\item{Lat}{Value or column name in data set @ForestGeoSubset (none) corresponding to Tree latitude}
\item{Genus}{Value or column name in data set @ForestGeoSubset (Genus) corresponding to Genus name}
\item{SubPlot}{Value or column name in data set @ForestGeoSubset (quadrat) corresponding to SubPlot name}
\item{DiameterUnitMan}{Value or column name in data set @ForestGeoSubset (mm) corresponding to What is your DBH unit ?}
\item{Xplot}{Value or column name in data set @ForestGeoSubset (gx) corresponding to Tree X euclidean position in plot}
\item{Xsubplot}{Value or column name in data set @ForestGeoSubset (none) corresponding to Tree X euclidean position in SUBplot}
\item{plotUnitMan}{Value or column name in data set @ForestGeoSubset (m) corresponding to XY plot coordinates units}
\item{subplotUnitMan}{Value or column name in data set @ForestGeoSubset (none) corresponding to XY subplot coordinates units}
\item{Diameter}{Value or column name in data set @ForestGeoSubset (dbh) corresponding to DBH}
\item{HOMUnitMan}{Value or column name in data set @ForestGeoSubset (m) corresponding to What is your HOM unit?}
\item{Variety}{Value or column name in data set @ForestGeoSubset (none) corresponding to Variety name}
\item{Voucher}{Value or column name in data set @ForestGeoSubset (Voucher) corresponding to Voucher code}
\item{Tidy}{Some value entered via interaction with the Shiny app}
\item{ClearValueName}{Some value entered via interaction with the Shiny app}
\item{TickedMelt3}{Some value entered via interaction with the Shiny app}
\item{Variablecolumns1}{Some value entered via interaction with the Shiny app}
\item{ValueName4}{Some value entered via interaction with the Shiny app}
\item{TickedMelt1}{Some value entered via interaction with the Shiny app}
\item{TickedMelt4}{Some value entered via interaction with the Shiny app}
\item{Variablecolumns3}{Some value entered via interaction with the Shiny app}
\item{ValueName3}{Some value entered via interaction with the Shiny app}
\item{TickedMelt2}{Some value entered via interaction with the Shiny app}
\item{Variablecolumns2}{Some value entered via interaction with the Shiny app}
\item{ValueName2}{Some value entered via interaction with the Shiny app}
\item{Variablecolumns4}{Some value entered via interaction with the Shiny app}
\item{ValueName1}{Some value entered via interaction with the Shiny app}
...
}
}
\usage{
ForestGeoProfile
}
\description{
Generated by interacting with app or function RequiredFormat_interactive
}
\keyword{datasets}
