% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParacouProfile.R
\docType{data}
\name{ParacouProfile}
\alias{ParacouProfile}
\title{ParacouProfile}
\format{
A named list
\describe{
\item{Site}{The column name in datda set @ParacouSubset (Forest) corresponding to Site name}
\item{Plot}{The column name in datda set @ParacouSubset (Plot) corresponding to Plot name}
\item{SubPlot}{The column name in datda set @ParacouSubset (SubPlot) corresponding to SubPlot name}
\item{CensusYear}{The column name in datda set @ParacouSubset (CensusYear) corresponding to Census year}
\item{CensusDate}{The column name in datda set @ParacouSubset (none) corresponding to Census Date}
\item{TreeFieldNum}{The column name in datda set @ParacouSubset (TreeFieldNum) corresponding to Tree unique identifiers}
\item{IdTree}{The column name in datda set @ParacouSubset (idTree) corresponding to Tree ID}
\item{IdStem}{The column name in datda set @ParacouSubset (none) corresponding to Stem ID}
\item{IdMeasure}{The column name in datda set @ParacouSubset (none) corresponding to Measure ID}
\item{LifeStatus}{The column name in datda set @ParacouSubset (CodeAlive) corresponding to Life status}
\item{DBH}{The column name in datda set @ParacouSubset (none) corresponding to DBH}
\item{DBHUnit}{The column name in datda set @ParacouSubset (none) corresponding to DBH unit}
\item{Circ}{The column name in datda set @ParacouSubset (Circ) corresponding to Circumference}
\item{CircUnit}{The column name in datda set @ParacouSubset (none) corresponding to Circumference unit}
\item{POM}{The column name in datda set @ParacouSubset (none) corresponding to POM}
\item{POMUnit}{The column name in datda set @ParacouSubset (none) corresponding to POM units}
\item{Lat}{The column name in datda set @ParacouSubset (Lat) corresponding to Tree latitude}
\item{Lon}{The column name in datda set @ParacouSubset (Lon) corresponding to Tree longitude}
\item{Xutm}{The column name in datda set @ParacouSubset (Xutm) corresponding to Tree X UTM coordinate}
\item{Yutm}{The column name in datda set @ParacouSubset (Yutm) corresponding to Tree Y UTM coordinate}
\item{Xfield}{The column name in datda set @ParacouSubset (Xfield) corresponding to Tree X euclidean position in plot}
\item{Yfield}{The column name in datda set @ParacouSubset (Yfield) corresponding to Tree Y euclidean position in plot}
\item{ScientificName}{The column name in datda set @ParacouSubset (none) corresponding to Scientific name}
\item{VernName}{The column name in datda set @ParacouSubset (VernName) corresponding to Vernacular name}
\item{Family}{The column name in datda set @ParacouSubset (Family) corresponding to Family name}
\item{Genus}{The column name in datda set @ParacouSubset (Genus) corresponding to Genus name}
\item{Species}{The column name in datda set @ParacouSubset (Species) corresponding to Species name}
\item{CommercialSp}{The column name in datda set @ParacouSubset (CommercialSp) corresponding to Species considered commercial}
\item{TreeHeight}{The column name in datda set @ParacouSubset (none) corresponding to Tree height}
\item{TreeHeightUnit}{The column name in datda set @ParacouSubset (none) corresponding to Tree height units}
\item{LifeForm}{The column name in datda set @ParacouSubset (none) corresponding to Life form}
\item{Treatment}{The column name in datda set @ParacouSubset (none) corresponding to Logging treatment}
\item{SoilType}{The column name in datda set @ParacouSubset (none) corresponding to Type of soil}
\item{Topography}{The column name in datda set @ParacouSubset (none) corresponding to Topographical position (lowland, plateau)}
\item{CircUnitMan}{The column name in datda set @ParacouSubset (cm) corresponding to What is your circumference unit?}
...
}
}
\usage{
ParacouProfile
}
\description{
Generated by interacting with function RequiredFormat_interactive
}
\keyword{datasets}
