% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParacouProfile.R
\docType{data}
\name{ParacouProfile}
\alias{ParacouProfile}
\title{ParacouProfile}
\format{
A named list
\describe{
\item{Site}{Value or column name in data set @ParacouSubset (Forest) corresponding to Site name}
\item{Plot}{Value or column name in data set @ParacouSubset (Plot) corresponding to Plot name}
\item{SubPlot}{Value or column name in data set @ParacouSubset (SubPlot) corresponding to SubPlot name}
\item{PlotArea}{Value or column name in data set @ParacouSubset (PlotArea) corresponding to Plot (or subplot) area}
\item{Year}{Value or column name in data set @ParacouSubset (Year) corresponding to Census year}
\item{Date}{Value or column name in data set @ParacouSubset (Date) corresponding to Census Date}
\item{TreeFieldNum}{Value or column name in data set @ParacouSubset (TreeFieldNum) corresponding to Tree unique identifiers}
\item{IdTree}{Value or column name in data set @ParacouSubset (idTree) corresponding to Tree ID}
\item{IdStem}{Value or column name in data set @ParacouSubset (none) corresponding to Stem ID}
\item{IdMeasure}{Value or column name in data set @ParacouSubset (none) corresponding to Measure ID}
\item{LifeStatus}{Value or column name in data set @ParacouSubset (CodeAlive) corresponding to Life status}
\item{DBH}{Value or column name in data set @ParacouSubset (none) corresponding to DBH}
\item{DBHUnit}{Value or column name in data set @ParacouSubset (none) corresponding to DBH unit}
\item{Circ}{Value or column name in data set @ParacouSubset (Circ) corresponding to Circumference}
\item{CircUnit}{Value or column name in data set @ParacouSubset (none) corresponding to Circumference unit}
\item{POM}{Value or column name in data set @ParacouSubset (none) corresponding to POM}
\item{POMUnit}{Value or column name in data set @ParacouSubset (mm) corresponding to POM units}
\item{Lat}{Value or column name in data set @ParacouSubset (Lat) corresponding to Tree latitude}
\item{Lon}{Value or column name in data set @ParacouSubset (Lon) corresponding to Tree longitude}
\item{Xutm}{Value or column name in data set @ParacouSubset (Xutm) corresponding to Tree X UTM coordinate}
\item{Yutm}{Value or column name in data set @ParacouSubset (Yutm) corresponding to Tree Y UTM coordinate}
\item{Xfield}{Value or column name in data set @ParacouSubset (Xfield) corresponding to Tree X euclidean position in plot}
\item{Yfield}{Value or column name in data set @ParacouSubset (Yfield) corresponding to Tree Y euclidean position in plot}
\item{ScientificName}{Value or column name in data set @ParacouSubset (none) corresponding to Scientific name}
\item{VernName}{Value or column name in data set @ParacouSubset (VernName) corresponding to Vernacular name}
\item{Family}{Value or column name in data set @ParacouSubset (Family) corresponding to Family name}
\item{Genus}{Value or column name in data set @ParacouSubset (Genus) corresponding to Genus name}
\item{Species}{Value or column name in data set @ParacouSubset (Species) corresponding to Species name}
\item{CommercialSp}{Value or column name in data set @ParacouSubset (CommercialSp) corresponding to Commercial species indicator}
\item{TreeHeight}{Value or column name in data set @ParacouSubset (none) corresponding to Tree height}
\item{TreeHeightUnit}{Value or column name in data set @ParacouSubset (mm) corresponding to Tree height units}
\item{LifeForm}{Value or column name in data set @ParacouSubset (none) corresponding to Life form}
\item{Treatment}{Value or column name in data set @ParacouSubset (none) corresponding to Logging treatment}
\item{SoilType}{Value or column name in data set @ParacouSubset (none) corresponding to Type of soil}
\item{Topography}{Value or column name in data set @ParacouSubset (none) corresponding to Topographical position (lowland, plateau)}
\item{CircUnitMan}{Value or column name in data set @ParacouSubset (cm) corresponding to What is your circumference unit?}
\item{PlotAreaUnitMan}{Value or column name in data set @ParacouSubset (ha) corresponding to What the unit for plot area?}
\item{DateFormat}{Value or column name in data set @ParacouSubset (\%Y-\%m-\%d) corresponding to Census Date format}
\item{IsLive}{Value or column name in data set @ParacouSubset (TRUE) corresponding to Which of your status(es) represent a LIVE tree?}
\item{IsCommercrial}{Value or column name in data set @ParacouSubset (TRUE) corresponding to Which of the following(s) indicate a commercial species?}
...
}
}
\usage{
ParacouProfile
}
\description{
Generated by interacting with app or function RequiredFormat_interactive
}
\keyword{datasets}
