% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParacouSubsetFormated.R
\docType{data}
\name{ParacouSubsetFormated}
\alias{ParacouSubsetFormated}
\title{ParacouSubsetFormated}
\format{
A tibble
\describe{
\item{Site}{Site level name}
\item{Plot}{Plot level name}
\item{SubPlot}{SubPlot level name}
\item{PlotArea}{Size of the plot in ha}
\item{SubPlotArea}{Size of the plot in ha}
\item{IdCensus}{Year of measurement}
\item{Year}{Year of measurement}
\item{Date, DateOriginal}{Date of measurement}
\item{TreeFieldNum}{Tree unique identifiers, matching the tag number in the field if that column was provided in the input, automatically generated if not.}
\item{IdTree}{Tree unique identifiers. This was automatically generated if not provided in the input data}
\item{LifeStatus, LifeStatusOriginal}{Logical: TRUE is tree was alive, FALSE if it was dead}
\item{DBH}{Diameter at breast height in cm}
\item{Circ}{circumference in cm}
\item{POM}{Height of measurement in m}
\item{Lat}{Latitude in degrees}
\item{Lon}{Longitude in degrees}
\item{Xutm}{X UTM coordinates in m}
\item{Yutm}{Y UTM coordinats in m}
\item{Xfield}{Tree X euclidean position in plot in m}
\item{Yfield}{Tree Y euclidean position in plot in m}
\item{ScientificName}{Scientific name}
\item{VernName}{Vernacular name}
\item{Family}{Family name}
\item{Genus}{Genus name}
\item{Species}{Species name}
\item{IdLevel}{Logical: TRUE if species is considered commercial, FALSE if not}
\item{Authority}{Logical: TRUE if species is considered commercial, FALSE if not}
\item{CommercialSp, CommercialSpOriginal}{Logical: TRUE if species is considered commercial, FALSE if not}
\item{TreeHeight}{Tree Height in m}
\item{LifeForm}{Life form}
...
}
}
\usage{
ParacouSubsetFormated
}
\description{
This is the standardized, formated version of ParacouSubset (see help(ParacouSubset))
}
\keyword{datasets}
