% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhylogeneticHierarchicalCorrection.R
\name{PhylogeneticHierarchicalCorrection}
\alias{PhylogeneticHierarchicalCorrection}
\title{Phylogenetic Hierarchical Correction}
\usage{
PhylogeneticHierarchicalCorrection(
  DataTree,
  Data,
  cresc,
  cresc_abs,
  cresc_abn,
  DBHCor,
  Time,
  PositiveGrowthThreshold,
  NegativeGrowthThreshold,
  DBHRange = 10,
  MinIndividualNbr = 5
)
}
\arguments{
\item{DataTree}{A dataset corresponding to a single tree's (1 IdTree)
measurements (data.table)
The dataset must contain the columns:
\itemize{
\item 'IdTree' (character)
\item 'ScientificName' (character)
\item 'Genus' (character)
\item 'Family' (character)
\item 'Diameter' (numeric)
\item 'Year' (numeric)
}}

\item{Data}{Complete dataset (data.table)
The dataset must contain the columns:
\itemize{
\item 'IdTree' (character)
\item 'ScientificName' (character)
\item 'Genus' (character)
\item 'Family' (character)
\item 'Diameter' (numeric)
\item 'Year' (numeric)
}}

\item{cresc}{Annual diameter increment (numeric)}

\item{cresc_abs}{Absolute diameter increment (not divided by time between 2
values) (numeric)}

\item{cresc_abn}{Abnormal diameter increment positions (numeric)}

\item{DBHCor}{Diameter vector in cm (numeric)}

\item{Time}{Time variable in years (numeric)}

\item{PositiveGrowthThreshold}{in cm/year : a tree
widening by more than x cm/year is considered abnormal (numeric, 1 value)}

\item{NegativeGrowthThreshold}{in cm/census : The possible
positive measurement error (+n) cannot be corrected until the growth
appears abnormal, but a negative measurement error can be allowed until -n
(a tree does not decrease). Thus the positive measurement error (+n) is
"compensated". (numeric, 1 value)}

\item{DBHRange}{DBH range in cm to take into account to select other trees in
the dataset to apply "phylogenetic hierarchical" correction (Default: 10
cm) (numeric, 1 value)}

\item{MinIndividualNbr}{Minimum number of individuals to take into account in
"phylogenetic hierarchical" correction (Default: 5) (numeric, 1 value)}
}
\value{
Fill columns:
\itemize{
\item \emph{DBHCor}: corrected trees diameter at default HOM
\item \emph{DiameterCorrectionMeth} = "species"/"genus"/"family"/"stand"/"shift
realignment"
}
}
\description{
Corrects a series of abnormal diameters of the tree using the
mean growth of other trees (with a minimum number of trees) of the same
species/genus/family/stand of the same diameter class.
}
\details{
It is assumed that these abnormal diameters were measured at a
different height than the default. We therefore correct the 1st value
with the mean growth of other trees. The following values are
corrected from the corrected value of the 1st value, keeping the
originally measured growth.
\itemize{
\item Correct the 1st value in the series:
\itemize{
\item Estimate the value that the 1st abnormal DBH should have: Estimated DBH =
previous DBH value + estimated growth by regression interpolation
\item Find other individual of the the same diameter class (\emph{DBHRange}) and of
the same species, genus, family or stand than the tree to correct, until
the minimum number of individuals required to consider the correction is
reached (\emph{MinIndividualNbr})
\item 1st abnormal DBH in the series = previous value + mean growth of the
other trees
}
\item Correct the following values in the series: The other abnormal DBH in the
series = previous value + their original growth.
}
}
\examples{
data(TestData)
DataTree <- TestData[IdTree \%in\% "100658"]

# Inputs
DataTree$Year <-  c(2000, 2002, 2004, 2006, 2008, 2010)
DataTree$Diameter <- c(13, 14, 15, 12, 13, 14)
cresc <- c(0.5, 0.5, NA, 0.5, 0.5)
cresc_abs <- c(1, 1, NA, 1, 1)
cresc_abn <- 3


DataTree_output <- PhylogeneticHierarchicalCorrection(
    DataTree = DataTree,
    Data = TestData,
    cresc = cresc, cresc_abs = cresc_abs, cresc_abn = cresc_abn,
    DBHCor = DataTree$Diameter, Time = DataTree$Year,
    PositiveGrowthThreshold = 5,
    NegativeGrowthThreshold = -2,
    DBHRange = 10, MinIndividualNbr = 5
)
}
