% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecruitmentCorrection.R
\name{RecruitmentCorrection}
\alias{RecruitmentCorrection}
\title{RecruitmentCorrection}
\usage{
RecruitmentCorrection(
  Data,
  MinDBH = 10,
  PositiveGrowthThreshold = 5,
  InvariantColumns = c("Site", "Genus", "Species", "Family", "ScientificName"),
  DetectOnly = FALSE
)
}
\arguments{
\item{Data}{(data.frame or data.table)}

\item{MinDBH}{Minimum diameter of trees inventoried according to your
protocol (in cm) (numeric, 1 value)}

\item{PositiveGrowthThreshold}{A tree widening by more than x
cm/year is considered abnormal (numeric, 1 value)}

\item{InvariantColumns}{Vector with the names of the columns that are
supposed to have always the same value for each measurement of the same
tree (character)}

\item{DetectOnly}{TRUE: Only detect errors, FALSE: detect and correct errors
(logical)}
}
\value{
Add rows for forgotten recruits with them estimated DBH in the
'DBHCor' column, create a 'CorrectedRecruit' col (logical) and fill the
'Comment' column : "This DBH is/was the 1st recorded for this tree,
according to its annual growth and the census done for this plot, it should
have been recruited earlier according to your protocol (MinDBH)."
}
\description{
RecruitmentCorrection
}
\details{
If the size of the tree has never changed, or if there is only one
value the same value is kept for the added forgotten recruits.
}
\examples{
library(data.table)
data(TestData)
setnames(TestData, "DBH", "DBHCor")

Rslt <- RecruitmentCorrection(TestData)
IdCorr <- Rslt[CorrectedRecruit == TRUE, IdTree]
TreesCorr <- Rslt[IdTree \%in\% IdCorr]

library(ggplot2)
ggplot(TreesCorr) +
aes(x = CensusYear, y = DBHCor) +
  geom_line(size = 0.5, colour = "#112446") +
  geom_point(shape = "circle", size = 1.5, mapping = aes(color = CorrectedRecruit)) +
  theme_minimal() +
  facet_wrap(vars(IdTree), scales = "free")

}
