% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecruitmentCorrection.R
\name{RecruitmentCorrectionByTree}
\alias{RecruitmentCorrectionByTree}
\title{RecruitmentCorrectionByTree}
\usage{
RecruitmentCorrectionByTree(
  DataTree,
  MinDBH = 10,
  PositiveGrowthThreshold = 5,
  InvariantColumns = c("Site", "Genus", "Species", "Family", "ScientificName"),
  PlotCensuses,
  DetectOnly = FALSE
)
}
\arguments{
\item{DataTree}{A dataset corresponding to a single tree's (1 IdTree)
measurements (data.frame or data.table). The dataset should preferably
contain the column of corrected diameters: 'DBHCor', otherwise the function
will take the column "DBH"}

\item{MinDBH}{Minimum diameter of trees inventoried according to your
protocol (in cm) (numeric, 1 value)}

\item{PositiveGrowthThreshold}{A tree widening by more than x
cm/year is considered abnormal (numeric, 1 value)}

\item{InvariantColumns}{Vector with the names of the columns that are
supposed to have always the same value for each measurement of the same
tree (character)}

\item{PlotCensuses}{Census years for the plot in which the tree is (integer)}

\item{DetectOnly}{TRUE: Only detect errors, FALSE: detect and correct errors
(logical)}
}
\value{
Add rows for forgotten recruits with them estimated DBH in the
'DBHCor' column, create a 'CorrectedRecruit' col (logical) and fill the
'Comment' column : "This DBH is/was the 1st recorded for this tree,
according to its annual growth and the census done for this plot, it should
have been recruited earlier according to your protocol (MinDBH)."
}
\description{
RecruitmentCorrectionByTree
}
\details{
If the size of the tree has never changed, or if there is only one
value the same value is kept for the added forgotten recruits.
If the DBH has not been corrected (DBCor column does not exist), the
function will create it for the forgotten recruits. It is strongly
recommended to correct the DBH before correcting the recruits
}
\examples{
library(data.table)
DataTree <- data.table(IdTree = "a",
                       Year = seq(2000,2008, by = 2), # 2 years/census
                       DBHCor  = as.numeric(c(13:17)), # 1cm/census(0.5 cm/year)
                       Site = "Imaginary forest"
)
# 1st DBH = 13 > MinDBH. In 1998 the tree was 12cm, 11cm in 1996 and 10cm in
1994.
Rslt <- RecruitmentCorrectionByTree(DataTree,
                                    InvariantColumns = "Site",
                                    PlotCensuses = seq(1996,2016, by = 2))

}
