% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegressionInterpolation.R
\name{RegressionInterpolation}
\alias{RegressionInterpolation}
\title{RegressionInterpolation}
\usage{
RegressionInterpolation(Y, X, CorrectionType = "quadratic")
}
\arguments{
\item{Y}{dfghj}

\item{X}{sdf}

\item{CorrectionType}{sdfg}
}
\value{
qsdf
}
\description{
RegressionInterpolation
}
