% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RequiredFormat.R
\name{RequiredFormat}
\alias{RequiredFormat}
\title{RequiredFormat}
\usage{
RequiredFormat(
  Data,
  Plot = NULL,
  SubPlot = NULL,
  Time = NULL,
  IdTree = NULL,
  LifeStatus = NULL,
  Size = NULL,
  SizeUnit = NULL,
  POM = NULL,
  POMUnit = NULL,
  PlotArea = NULL,
  X = NULL,
  Y = NULL,
  ScientificName = NULL,
  VernName = NULL,
  Family = NULL,
  Genus = NULL,
  Species = NULL,
  CommercialSp = NULL,
  TreeHeight = NULL,
  TreeHeightUnit = NULL
)
}
\arguments{
\item{Data}{Input forest inventory (data.frame)}

\item{Plot}{The name of the column containing plot indices (character)}

\item{SubPlot}{The name of the column containing sub-plot indices (character)}

\item{Time}{The name of the column containing census years (character)}

\item{IdTree}{The name of the column containing tree unique identifiers
(character)}

\item{LifeStatus}{The name of the column containing tree vital status
(character)}

\item{Size}{The name of the column containing tree size measurements,
diameter or circumference (character)}

\item{SizeUnit}{The 'Size' measurement unit of the tree. The possible values
are 'mm' for millimetres, 'cm' for centimetres, 'dm' for decimetres and 'm'
for metres"(character)}

\item{POM}{The name of the column containing Point of Measurement (POM)
values: height at which the 'Size' was measured, or POM code (character)}

\item{POMUnit}{The 'POM' unit of the tree. The possible values are 'mm' for
millimetres, 'cm' for centimetres, 'dm' for decimetres and 'm' for metres"
(character)}

\item{PlotArea}{The name of the column containing the area of each plot, in
hectares (character)}

\item{X}{The name of the column containing the tree X coordinate (character)}

\item{Y}{The name of the column containing the tree Y coordinate (character)}

\item{ScientificName}{The name of the column containing the binary scientific
name of the tree (character)}

\item{VernName}{The name of the column containing the vernacular name of the
tree (character)}

\item{Family}{The name of the column containing the family name of the tree
(character)}

\item{Genus}{The name of the column containing the genus name of the tree
(character)}

\item{Species}{The name of the column containing the species name of the tree
(character)}

\item{CommercialSp}{The name of the column containing the commercial or
non-commercial nature of the tree species (character)}

\item{TreeHeight}{The name of the column containing the the measured height
of the tree (character)}

\item{TreeHeightUnit}{The 'TreeHeight' unit. The possible values are 'mm' for
millimetres, 'cm' for centimetres, 'dm' for decimetres and 'm' for
metres"(character)}
}
\value{
Input inventory (data.frame) in the required package format.
}
\description{
RequiredFormat
}
\examples{
\dontrun{
RequiredFormat(Data,
               Plot = NULL,
               SubPlot = NULL,
               Time = NULL,
               IdTree = NULL,
               LifeStatus = NULL,
               Size = NULL,
               SizeUnit = NULL,
               POM = NULL,
               POMUnit = NULL,
               PlotArea = NULL,
               X = NULL,
               Y = NULL,
               ScientificName = NULL,
               VernName = NULL,
               Family = NULL,
               Genus = NULL,
               Species = NULL,
               CommercialSp = NULL,
               TreeHeight = NULL,
               TreeHeightUnit = NULL)
               }

}
