% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatusCorrection.R
\name{StatusCorrection}
\alias{StatusCorrection}
\title{StatusCorrection}
\usage{
StatusCorrection(
  Data,
  InvariantColumns = c("Site", "Genus", "Species", "Family", "ScientificName"),
  DeathConfirmation = 2,
  UseSize = FALSE,
  DetectOnly = FALSE,
  RemoveRBeforeAlive = FALSE,
  RemoveRAfterDeath = FALSE
)
}
\arguments{
\item{Data}{(data.frame or data.table)
The \emph{LifeStatus} column must be coded as:
- TRUE = alive,
- FALSE = dead,
- NA = unseen
The \emph{Plot} column is needed to add rows to the census where the plot was
inventoried, where the tree was alive, but not recorded.}

\item{InvariantColumns}{Vector with the names of the columns that are
supposed to have always the same value for each measurement of the same
tree (character)}

\item{DeathConfirmation}{Number of times (censuses) needed for an unseen tree
to be considered dead (numeric)  (Default = 2 censuses)}

\item{UseSize}{Use the size presence as a witness of the living status of the
tree (logical) (Default = FALSE)}

\item{DetectOnly}{TRUE: Only detect errors, FALSE: detect and correct errors
(logical)}

\item{RemoveRBeforeAlive}{Do you want to delete the rows about the tree
before it was seen alive for the 1st time? (logical)}

\item{RemoveRAfterDeath}{After correction do you want to delete the rows
concerning the tree after its death? (logical)}
}
\value{
Fill the \emph{Comment} column with error type informations. If
\emph{DetectOnly} = FALSE, add a \emph{LifeStatusCor} column with the corrected trees
life status.
}
\description{
Detect errors or detect errors and correct the tree life status
evolution over the censuses.
Inspired by the code of Nino Page package (ForestData::correct_alive() and
.correct_alive_tree())
}
\details{
\itemize{
\item if UseSize : if DBH != NA -> Alive
\item \emph{Dead} > Alive -> NA
\item add rows for the forgotten censuses between 2 'Alive'
\item Alive > \emph{Dead}/\emph{NA} > Alive -> Alive
\item Alive > \emph{NA} > Dead -> NA
\item Alive > \emph{Dead} > NA -> Dead
\item Alive > \emph{NA} > \emph{NA}:
if DeathConfirmation > unseens -> NA
if DeathConfirmation =< unseens -> Dead
}
}
\examples{
library(data.table)
data(TestData)

selection <- c("101184", "101433","101435","101436")

# Write the sequence
TestData <- TestData[order(Year)] # arrange years in ascending order
TestData[IdTree == "101184", LifeStatus := c(TRUE, TRUE, TRUE, TRUE, FALSE)]
TestData[IdTree == "101433", LifeStatus := c(FALSE, TRUE, TRUE, TRUE, TRUE)]
TestData[IdTree == "101435", LifeStatus := c(TRUE, TRUE, NA, FALSE, TRUE)]
TestData[IdTree == "101436", LifeStatus := c(TRUE, NA, NA, FALSE, NA)]


Rslt <- StatusCorrection(TestData[IdTree \%in\% selection])

library(ggplot2)
ggplot(Rslt) +
aes(x = Year, y = LifeStatusCor) +
  geom_point(shape = "circle", size = 3.9, colour = "#46337E") +
  geom_smooth(span = 0.75) +
  theme_minimal() +
  facet_wrap(vars(IdTree), scales = "free")

}
