% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaperCorrection.R
\name{TaperCorrection}
\alias{TaperCorrection}
\title{TaperCorrection}
\usage{
TaperCorrection(
  Data,
  DefaultHOM = 1.3,
  TaperParameter = function(DAB, HOM) 0.156 - 0.023 * log(DAB) - 0.021 * log(HOM),
  TaperFormula = function(DAB, HOM, TaperParameter, DefaultHOM = DefaultHOM)
    DAB/(exp(-TaperParameter * (HOM - DefaultHOM))),
  DetectOnly = FALSE
)
}
\arguments{
\item{Data}{Dataset (data.frame or data.table)
The dataset must contain the columns:
\itemize{
\item \code{Diameter} (numeric)
\item \strong{\code{HOM} (Height Of Measurement) (numeric)}
}}

\item{DefaultHOM}{Default Height Of Measurement in meter (Default: 1.3 m)
(numeric, 1 value)}

\item{TaperParameter}{Taper parameter (unitless) formula (function)
Default: \emph{TaperParameter = 0.156 - 0.023 log(DAB) - 0.021 log(HOM)}
of Cushman et al.2021.
With:
\itemize{
\item \emph{DAB}: Diameter Above Buttress (in cm)
\item \emph{HOM}: Height Of Measurement (in m)
}}

\item{TaperFormula}{Taper formula (function)
Default: \emph{DAB / (e^(- TaperParameter (HOM - DefaultHOM)))}
of Cushman et al.2021.
With:
\itemize{
\item \emph{DAB}: Diameter Above Buttress (in cm)
\item \emph{HOM}: Height Of Measurement (in m)
\item \emph{DefaultHOM}:  Default Height Of Measurement (in m)
\item \emph{TaperParameter}: Taper parameter (unitless)
}}

\item{DetectOnly}{TRUE: Only detect errors, FALSE: detect and correct errors
(Default: FALSE) (logical)}
}
\value{
Fill the \emph{Comment} column with error type informations. If
\emph{DetectOnly} = FALSE, add columns:
- \emph{TaperCorDBH}: corrected trees diameter at default HOM
- \emph{DiameterCorrectionMeth} = "taper"
- \emph{HOMCor} = HOM corresponding to the \emph{TaperCorDBH} (= \emph{DefaultHOM})
}
\description{
TaperCorrection
}
\examples{
library(data.table)

Data <- data.table(
      Diameter = c(13:16, 16-4, (16-4)+2, (16-4)+3, 15-4, (15-4)+2),
      POM = c(0, 0, 0, 0, 1, 1, 1, 2, 2),
      HOM = c(1.3, 1.3, 1.3, 1.3, 1.5, 1.5, 1.5, 2, 2))

Rslt <- TaperCorrection(Data)

}
